refactor(personal-assistant): collapse mixin chain into concrete LifeOpsService facade
Phase C — eliminate the mixin SYSTEM. service.ts is now a single concrete
`class LifeOpsService extends LifeOpsServiceBase` holding all 26 domain fields
(deps wired via typed `this` arrows, no casts), 263 flat delegators (preserving
the existing call-site surface), and 26 typed namespace getters (service.goals,
service.gmail, ... for ergonomic new-code access). Deleted: the 26 `withX` mixin
functions + their `as unknown as MixinClass<>` return casts, every dependency
`(this as unknown as XDeps)` callback cast, the `Constructor<StatusMixinDependencies>`
cast, the `LIFEOPS_BASE..LIFEOPS_COMPOSED` chain, and the
noUnsafeDeclarationMerging interface-merge with its 4 Omit<> collision deconflicts
(a concrete class surfaces every method directly). Mixin files keep only their
exported LifeOps<D>Service interfaces.

Result: lifeops `as unknown as` 79 -> 5 (assembly 76 -> 2 genuine cross-package
type-drift bridges; domains cast-free). Typecheck 0 errors in scope; vitest
16 files / 112 tests pass (the prior bun-runner '4 fails' were runner artifacts —
vitest runs them green). No behaviour change; the depth-limit and collision
problems that forced the casts are gone with the chain.
