+++ plugins/plugin-personal-assistant/src/lifeops/service-mixin-status.ts
@@ -1,10 +1,6 @@
 import type { LifeOpsCapabilitiesStatus } from "@elizaos/shared";
-import { type StatusDeps, StatusDomain } from "./domains/status-service.js";
-import type {
-  Constructor,
-  LifeOpsServiceBase,
-  MixinClass,
-} from "./service-mixin-core.js";
+import type { StatusDeps } from "./domains/status-service.js";
+import type { LifeOpsServiceBase } from "./service-mixin-core.js";
 
 export interface LifeOpsStatusService {
   getCapabilityStatus(now?: Date): Promise<LifeOpsCapabilitiesStatus>;
@@ -16,34 +12,3 @@ export interface LifeOpsStatusService {
  * {@link StatusDeps}; this alias keeps the explicit composition-root cast typed.
  */
 export type StatusMixinDependencies = LifeOpsServiceBase & StatusDeps;
-
-/** @internal */
-export function withStatus<TBase extends Constructor<LifeOpsServiceBase>>(
-  Base: TBase,
-): MixinClass<TBase, LifeOpsStatusService> {
-  class LifeOpsStatusServiceMixin extends Base {
-    // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
-    // Public (not private) to avoid TS4094 on the re-exported mixin class.
-    readonly statusDomain = new StatusDomain(this, {
-      getScheduleMergedState: (...args) =>
-        (this as unknown as StatusDeps).getScheduleMergedState(...args),
-      getBrowserSettings: (...args) =>
-        (this as unknown as StatusDeps).getBrowserSettings(...args),
-      listBrowserCompanions: (...args) =>
-        (this as unknown as StatusDeps).listBrowserCompanions(...args),
-      getXConnectorStatus: (...args) =>
-        (this as unknown as StatusDeps).getXConnectorStatus(...args),
-      getHealthConnectorStatus: (...args) =>
-        (this as unknown as StatusDeps).getHealthConnectorStatus(...args),
-    });
-
-    getCapabilityStatus(now?: Date): Promise<LifeOpsCapabilitiesStatus> {
-      return this.statusDomain.getCapabilityStatus(now);
-    }
-  }
-
-  return LifeOpsStatusServiceMixin as unknown as MixinClass<
-    TBase,
-    LifeOpsStatusService
-  >;
-}
