+++ plugins/plugin-personal-assistant/src/lifeops/service.ts
@@ -1,177 +1,2574 @@
 /**
- * LifeOps Service — thin facade that composes domain-specific mixins.
+ * LifeOps Service — single concrete facade.
  *
- * The implementation lives in the `service-mixin-*.ts` files; standalone
- * helpers live in `service-normalize-*.ts` and `service-helpers-*.ts`.
- * This file only re-exports the public surface that consumers already import.
+ * Declares every domain sub-service field and delegates each public method to
+ * its domain. Standalone helpers live in `service-normalize-*.ts` and
+ * `service-helpers-*.ts`; the per-domain public interfaces and shared types are
+ * still exported from the `service-mixin-*.ts` files that consumers import.
  */
 
 export { LifeOpsServiceError } from "./service-types.js";
 
 import type {
-  LifeOpsReminderAttempt,
-  LifeOpsWorkflowRun,
+  BrowserBridgeCompanionAutoPairResponse,
+  BrowserBridgeCompanionPairingResponse,
+  BrowserBridgeCompanionRevokeResponse,
+  BrowserBridgeCompanionStatus,
+  BrowserBridgeCompanionSyncResponse,
+  BrowserBridgePageContext,
+  BrowserBridgeSettings,
+  BrowserBridgeTabSummary,
+  CreateBrowserBridgeCompanionAutoPairRequest,
+  CreateBrowserBridgeCompanionPairingRequest,
+  SyncBrowserBridgeStateRequest,
+  UpdateBrowserBridgeSettingsRequest,
+} from "@elizaos/plugin-browser";
+import type { DiscordMessageSearchResult } from "@elizaos/plugin-discord/user-account-scraper";
+import type {
+  DuffelOffer,
+  DuffelOrder,
+  DuffelPayment,
+  SearchFlightsRequest,
+  SearchFlightsResult,
+} from "@elizaos/plugin-elizacloud/cloud/duffel-client";
+import type { LifeOpsSubscriptionPlaybook } from "@elizaos/plugin-finances/subscriptions-playbooks";
+import type {
+  LifeOpsSubscriptionAuditSummary,
+  LifeOpsSubscriptionCancellationRequest,
+  LifeOpsSubscriptionCancellationSummary,
+  LifeOpsSubscriptionDiscoveryRequest,
+  LifeOpsSubscriptionExecutor,
+} from "@elizaos/plugin-finances/subscriptions-types";
+import type { GoogleDriveFile } from "@elizaos/plugin-google";
+import type {
+  HealthBackend,
+  HealthDailySummary,
+  HealthDataPoint,
+  LifeOpsDerivedEvent,
+  ScreenTimeAggregateRow,
+  ScreenTimeWeeklyAverageItem,
+} from "@elizaos/plugin-health";
+import type {
+  EmailSubscriptionScanResult,
+  EmailUnsubscribeRecord,
+  EmailUnsubscribeRequest,
+  EmailUnsubscribeResult,
+  EmailUnsubscribeScanRequest,
+} from "@elizaos/plugin-inbox/inbox/email-unsubscribe-types";
+import type {
+  CreateLifeOpsCalendarEventAttendee,
+  CreateLifeOpsCalendarEventRequest,
+  GetLifeOpsCalendarFeedRequest,
+  GetLifeOpsInboxRequest,
+  LifeOpsCalendarEvent,
+  LifeOpsCalendarFeed,
+  LifeOpsCalendarSummary,
+  LifeOpsCapabilitiesStatus,
+  LifeOpsDiscordConnectorStatus,
+  LifeOpsIMessageConnectorStatus,
+  LifeOpsInbox,
+  LifeOpsInboxMessage,
+  LifeOpsMessageChannel,
+  LifeOpsNextCalendarEventContext,
+  LifeOpsOwnerBrowserAccessSource,
+  LifeOpsPersonalBaselineResponse,
+  LifeOpsRelationship,
+  LifeOpsRelationshipInteraction,
+  LifeOpsSchedulingNegotiation,
+  LifeOpsSchedulingProposal,
+  LifeOpsScreenTimeDaily,
+  LifeOpsScreenTimeHistoryResponse,
+  LifeOpsScreenTimeRangeKey,
+  LifeOpsScreenTimeSession,
+  LifeOpsScreenTimeSource,
+  LifeOpsScreenTimeSummary,
+  LifeOpsSignalConnectorStatus,
+  LifeOpsSignalInboundMessage,
+  LifeOpsSleepHistoryResponse,
+  LifeOpsSleepRegularityResponse,
+  LifeOpsTelegramConnectorStatus,
+  LifeOpsWhatsAppConnectorStatus,
+  LifeOpsXFeedItem,
+  LifeOpsXFeedType,
+  ListLifeOpsCalendarsRequest,
+  LifeOpsScreenTimeBreakdown as ScreenTimeBreakdown,
+  LifeOpsSocialHabitSummary as SocialHabitSummary,
+  VerifyLifeOpsTelegramConnectorRequest,
+  VerifyLifeOpsTelegramConnectorResponse,
 } from "@elizaos/shared";
-import type { BrowserBridgeService } from "./service-mixin-browser.js";
-import { withBrowser } from "./service-mixin-browser.js";
-// Public method interfaces each mixin contributes. The composed class type
-// exceeds TypeScript's mixin-inference depth (~6 chained generics), so the
-// `interface LifeOpsService` merge below restates them explicitly to surface
-// every mixin method on the service type.
-import type { LifeOpsCalendarService } from "./service-mixin-calendar.js";
-import { withCalendar } from "./service-mixin-calendar.js";
-import type { Constructor } from "./service-mixin-core.js";
-import { LifeOpsServiceBase } from "./service-mixin-core.js";
-import type { LifeOpsDefinitionService } from "./service-mixin-definitions.js";
-import { withDefinitions } from "./service-mixin-definitions.js";
-import type { LifeOpsDiscordService } from "./service-mixin-discord.js";
-import { withDiscord } from "./service-mixin-discord.js";
-import type { LifeOpsDriveService } from "./service-mixin-drive.js";
-import { withDrive } from "./service-mixin-drive.js";
-import type { LifeOpsEmailUnsubscribeService } from "./service-mixin-email-unsubscribe.js";
-import { withEmailUnsubscribe } from "./service-mixin-email-unsubscribe.js";
-import type { LifeOpsGmailService } from "./service-mixin-gmail.js";
-import { withGmail } from "./service-mixin-gmail.js";
-import type { LifeOpsGoalService } from "./service-mixin-goals.js";
-import { withGoals } from "./service-mixin-goals.js";
-import type { LifeOpsGoogleService } from "./service-mixin-google.js";
-import { withGoogle } from "./service-mixin-google.js";
-import type { LifeOpsHealthServicePublic } from "./service-mixin-health.js";
-import { withHealth } from "./service-mixin-health.js";
-import type { LifeOpsIMessageService } from "./service-mixin-imessage.js";
-import { withIMessage } from "./service-mixin-imessage.js";
-import type { LifeOpsInboxService } from "./service-mixin-inbox.js";
-import { withInbox } from "./service-mixin-inbox.js";
-import type { LifeOpsRelationshipService } from "./service-mixin-relationships.js";
-import { withRelationships } from "./service-mixin-relationships.js";
-import type { LifeOpsReminderService } from "./service-mixin-reminders.js";
-import { withReminders } from "./service-mixin-reminders.js";
-import type { LifeOpsSchedulingService } from "./service-mixin-scheduling.js";
-import { withScheduling } from "./service-mixin-scheduling.js";
-import type { LifeOpsScreenTimeServicePublic } from "./service-mixin-screentime.js";
-import { withScreenTime } from "./service-mixin-screentime.js";
-import type { LifeOpsSignalService } from "./service-mixin-signal.js";
-import { withSignal } from "./service-mixin-signal.js";
-import { withSleep } from "./service-mixin-sleep.js";
-import type { LifeOpsStatusService } from "./service-mixin-status.js";
+import type {
+  CompleteLifeOpsBrowserSessionRequest,
+  CompleteLifeOpsOccurrenceRequest,
+  ConfirmLifeOpsBrowserSessionRequest,
+  CreateLifeOpsBrowserSessionRequest,
+  CreateLifeOpsDefinitionRequest,
+  CreateLifeOpsGmailBatchReplyDraftsRequest,
+  CreateLifeOpsGmailReplyDraftRequest,
+  CreateLifeOpsGoalRequest,
+  CreateLifeOpsWorkflowRequest,
+  CreateLifeOpsXPostRequest,
+  DisconnectLifeOpsGoogleConnectorRequest,
+  DisconnectLifeOpsHealthConnectorRequest,
+  GetLifeOpsGmailRecommendationsRequest,
+  GetLifeOpsGmailSearchRequest,
+  GetLifeOpsGmailSpamReviewRequest,
+  GetLifeOpsGmailTriageRequest,
+  GetLifeOpsGmailUnrespondedRequest,
+  GetLifeOpsHealthSummaryRequest,
+  IngestLifeOpsGmailEventRequest,
+  LifeOpsBrowserSession,
+  LifeOpsChannelPolicy,
+  LifeOpsConnectorGrant,
+  LifeOpsConnectorMode,
+  LifeOpsConnectorSide,
+  LifeOpsDefinitionRecord,
+  LifeOpsGmailBatchReplyDraftsFeed,
+  LifeOpsGmailBatchReplySendResult,
+  LifeOpsGmailEventIngestResult,
+  LifeOpsGmailManageResult,
+  LifeOpsGmailMessageSummary,
+  LifeOpsGmailNeedsResponseFeed,
+  LifeOpsGmailRecommendationsFeed,
+  LifeOpsGmailReplyDraft,
+  LifeOpsGmailSearchFeed,
+  LifeOpsGmailSpamReviewFeed,
+  LifeOpsGmailSpamReviewItem,
+  LifeOpsGmailTriageFeed,
+  LifeOpsGmailUnrespondedFeed,
+  LifeOpsGoalExperienceLoop,
+  LifeOpsGoalRecord,
+  LifeOpsGoalReview,
+  LifeOpsGoogleConnectorStatus,
+  LifeOpsHealthConnectorProvider,
+  LifeOpsHealthConnectorStatus,
+  LifeOpsHealthSummaryResponse,
+  LifeOpsOccurrenceExplanation,
+  LifeOpsOccurrenceView,
+  LifeOpsOverview,
+  LifeOpsWeeklyGoalReview,
+  LifeOpsWorkflowRecord,
+  LifeOpsWorkflowRun,
+  LifeOpsXConnectorStatus,
+  LifeOpsXDm,
+  LifeOpsXPostResponse,
+  ManageLifeOpsGmailMessagesRequest,
+  SendLifeOpsGmailBatchReplyRequest,
+  SendLifeOpsGmailMessageRequest,
+  SendLifeOpsGmailReplyRequest,
+  SnoozeLifeOpsOccurrenceRequest,
+  StartLifeOpsGoogleConnectorRequest,
+  StartLifeOpsGoogleConnectorResponse,
+  StartLifeOpsHealthConnectorRequest,
+  StartLifeOpsHealthConnectorResponse,
+  SyncLifeOpsHealthConnectorRequest,
+  UpdateLifeOpsBrowserSessionProgressRequest,
+  UpdateLifeOpsDefinitionRequest,
+  UpdateLifeOpsGmailSpamReviewItemRequest,
+  UpdateLifeOpsGoalRequest,
+  UpdateLifeOpsWorkflowRequest,
+} from "../contracts/index.js";
+import { BrowserDomain } from "./domains/browser-service.js";
+import { CalendarDomain } from "./domains/calendar-service.js";
+import { DefinitionsDomain } from "./domains/definitions-service.js";
 import {
-  type StatusMixinDependencies,
-  withStatus,
-} from "./service-mixin-status.js";
-import type { LifeOpsSubscriptionService } from "./service-mixin-subscriptions.js";
-import { withSubscriptions } from "./service-mixin-subscriptions.js";
-import type { LifeOpsTelegramService } from "./service-mixin-telegram.js";
-import { withTelegram } from "./service-mixin-telegram.js";
-import type { LifeOpsTravelServicePublic } from "./service-mixin-travel.js";
-import { withTravel } from "./service-mixin-travel.js";
-import type { LifeOpsWhatsAppService } from "./service-mixin-whatsapp.js";
-import { withWhatsApp } from "./service-mixin-whatsapp.js";
-import type { LifeOpsWorkflowService } from "./service-mixin-workflows.js";
-import { withWorkflows } from "./service-mixin-workflows.js";
-import type { LifeOpsXService } from "./service-mixin-x.js";
-import { withX } from "./service-mixin-x.js";
-import type { LifeOpsXReadService } from "./service-mixin-x-read.js";
-import { withXRead } from "./service-mixin-x-read.js";
+  type DiscordConnectorVerification,
+  DiscordDomain,
+  type DiscordSendMessageResult,
+} from "./domains/discord-service.js";
+import { DriveDomain } from "./domains/drive-service.js";
+import { EmailUnsubscribeDomain } from "./domains/email-unsubscribe-service.js";
+import { GmailDomain } from "./domains/gmail-service.js";
+import { GoalsDomain } from "./domains/goals-service.js";
+import { GoogleDomain } from "./domains/google-service.js";
+import { HealthDomain } from "./domains/health-service.js";
+import {
+  type IMessageChat,
+  type IMessageDeliveryResult,
+  IMessageDomain,
+  type IMessageRecord,
+  type IMessageSendRequest,
+} from "./domains/imessage-service.js";
+import { InboxDomain } from "./domains/inbox-service.js";
+import { RelationshipsDomain } from "./domains/relationships-service.js";
+import { RemindersDomain } from "./domains/reminders-service.js";
+import { SchedulingDomain } from "./domains/scheduling-service.js";
+import { ScreenTimeDomain } from "./domains/screentime-service.js";
+import { SignalDomain } from "./domains/signal-service.js";
+import { SleepDomain } from "./domains/sleep-service.js";
+import { StatusDomain } from "./domains/status-service.js";
+import { SubscriptionsDomain } from "./domains/subscriptions-service.js";
+import {
+  TelegramDomain,
+  type TelegramMessageSearchResult,
+  type TelegramReadReceiptResult,
+} from "./domains/telegram-service.js";
+import { TravelDomain } from "./domains/travel-service.js";
+import {
+  WhatsAppDomain,
+  type WhatsAppMessage,
+  type WhatsAppSendRequest,
+} from "./domains/whatsapp-service.js";
+import { WorkflowsDomain } from "./domains/workflows-service.js";
+import { XReadDomain } from "./domains/x-read-service.js";
+import { XDomain } from "./domains/x-service.js";
+import type {
+  LifeOpsScheduleInspection,
+  LifeOpsScheduleSummary,
+} from "./schedule-insight.js";
+import { LifeOpsServiceBase } from "./service-mixin-core.js";
+import type {
+  FlightBookingExecutionResult,
+  PreparedFlightBooking,
+  TravelBookingPassenger,
+  TravelCalendarSyncPlan,
+} from "./travel-booking.types.js";
+
+type ScreenTimeEventInput = {
+  source: "app" | "website";
+  identifier: string;
+  displayName: string;
+  startAt: string;
+  endAt?: string | null;
+  durationSeconds?: number;
+  metadata?: Record<string, unknown>;
+};
+
+type ScreenTimeWeeklyAverageResponse = {
+  items: ScreenTimeWeeklyAverageItem[];
+  totalSeconds: number;
+  daysInWindow: number;
+};
+
+type XReadOpts = {
+  limit?: number;
+};
+
+type XFeedReadOpts = XReadOpts & {
+  query?: string;
+};
+
+type OptionalXGrantResolver = {
+  resolveXGrant?: () => Promise<LifeOpsConnectorGrant | null>;
+};
 
 /**
- * Mixin order follows dependency direction: Google auth → data layers
- * (Calendar, Gmail, Drive) → business logic (Reminders, Browser, Workflows,
- * Definitions, Goals) → connectors (X, Telegram, Discord, Signal).
+ * Main LifeOps service — every domain sub-service composed on one concrete
+ * class built directly on {@link LifeOpsServiceBase}.
  */
-const LIFEOPS_BASE = withGoogle(LifeOpsServiceBase);
-const LIFEOPS_WITH_DATA = withDrive(withGmail(withCalendar(LIFEOPS_BASE)));
-const LIFEOPS_WITH_BUSINESS = withGoals(
-  withDefinitions(withWorkflows(withBrowser(withReminders(LIFEOPS_WITH_DATA)))),
-);
-const LIFEOPS_WITH_X = withX(LIFEOPS_WITH_BUSINESS);
-const LIFEOPS_WITH_RELATIONS = withRelationships(LIFEOPS_WITH_X);
-const LIFEOPS_WITH_DOMAIN = withEmailUnsubscribe(
-  withHealth(LIFEOPS_WITH_RELATIONS),
-);
-const LIFEOPS_WITH_X_READ = withXRead(LIFEOPS_WITH_DOMAIN);
-const LIFEOPS_WITH_CONNECTORS = withWhatsApp(
-  withSignal(withDiscord(withTelegram(withIMessage(LIFEOPS_WITH_X_READ)))),
-);
-const LIFEOPS_WITH_TRAVEL = withTravel(LIFEOPS_WITH_CONNECTORS);
-const LIFEOPS_WITH_SCHEDULING = withScheduling(LIFEOPS_WITH_TRAVEL);
-// Payment-source / transaction / spending logic moved to
-// @elizaos/plugin-finances (FinancesService). Subscription audit / cancellation
-// also moved there (SubscriptionsService), which reaches Gmail + the browser
-// bridge through runtime-service seams. LifeOpsService no longer implements
-// either back-end; the OWNER_FINANCES handler + the /api/lifeops/money/* and
-// /api/lifeops/subscriptions/* routes delegate to the finances services. The
-// `withSubscriptions` mixin is a thin forwarding shim that keeps the service
-// surface stable for those call sites.
-const LIFEOPS_WITH_SUBS = withSubscriptions(LIFEOPS_WITH_SCHEDULING);
-// TypeScript loses track of constraint satisfaction past ~6 chained generic
-// mixins, so we cast explicitly. The runtime composition has every method
-// `withStatus` depends on (getScheduleMergedState from withScheduling,
-// getBrowserSettings/listBrowserCompanions from withBrowser,
-// getXConnectorStatus from withX, getHealthConnectorStatus from withHealth).
-type LifeOpsSubsCtor = typeof LIFEOPS_WITH_SUBS;
-const LIFEOPS_WITH_STATUS = withStatus(
-  LIFEOPS_WITH_SUBS as LifeOpsSubsCtor & Constructor<StatusMixinDependencies>,
-);
-const LIFEOPS_COMPOSED = withInbox(
-  withSleep(withScreenTime(LIFEOPS_WITH_STATUS)),
-);
+export class LifeOpsService extends LifeOpsServiceBase {
+  get google() {
+    return this.googleDomain;
+  }
 
-class LifeOpsServiceComposed extends LIFEOPS_COMPOSED {}
+  get calendar() {
+    return this.calendarDomain;
+  }
 
-/**
- * Main LifeOps service — assembled from domain mixins layered on top of
- * {@link LifeOpsServiceBase}.
- */
-export class LifeOpsService extends LifeOpsServiceComposed {}
+  get gmail() {
+    return this.gmailDomain;
+  }
 
-/** Declared explicitly: mixin composition exceeds TypeScript inference depth. */
-// biome-ignore lint/suspicious/noUnsafeDeclarationMerging: intentional class+interface merge to surface mixin methods past TS inference depth
-export interface LifeOpsService
-  extends Omit<
-      BrowserBridgeService,
-      // both already surfaced (with slightly different signatures) by the
-      // composed class / status-mixin deps; omit to avoid an extends conflict.
-      "getBrowserSettings" | "listBrowserCompanions"
-    >,
-    LifeOpsSubscriptionService,
-    LifeOpsTelegramService,
-    LifeOpsIMessageService,
-    LifeOpsRelationshipService,
-    LifeOpsSignalService,
-    LifeOpsWhatsAppService,
-    LifeOpsCalendarService,
-    // `getHealthConnectorStatus` is also declared (with a different return
-    // type) by the status mixin already on the composed class, so omit it here.
-    Omit<LifeOpsHealthServicePublic, "getHealthConnectorStatus">,
-    LifeOpsDefinitionService,
-    LifeOpsDiscordService,
-    LifeOpsDriveService,
-    LifeOpsEmailUnsubscribeService,
-    LifeOpsGmailService,
-    LifeOpsGoalService,
-    LifeOpsGoogleService,
-    LifeOpsInboxService,
-    LifeOpsReminderService,
-    LifeOpsSchedulingService,
-    LifeOpsScreenTimeServicePublic,
-    LifeOpsStatusService,
-    LifeOpsTravelServicePublic,
-    LifeOpsWorkflowService,
-    LifeOpsXReadService,
-    // `getXConnectorStatus` is also surfaced (with a slightly different
-    // signature) by the status mixin already on the composed class, so omit it
-    // here to avoid an extends conflict while keeping the other X methods.
-    Omit<LifeOpsXService, "getXConnectorStatus"> {
-  processScheduledWork(request?: {
-    now?: string;
-    reminderLimit?: number;
-    workflowLimit?: number;
-    scheduledTaskLimit?: number;
+  get drive() {
+    return this.driveDomain;
+  }
+
+  get reminders() {
+    return this.remindersDomain;
+  }
+
+  get browser() {
+    return this.browserDomain;
+  }
+
+  get workflows() {
+    return this.workflowsDomain;
+  }
+
+  get definitions() {
+    return this.definitionsDomain;
+  }
+
+  get goals() {
+    return this.goalsDomain;
+  }
+
+  get x() {
+    return this.xDomain;
+  }
+
+  get relationships() {
+    return this.relationshipsDomain;
+  }
+
+  get emailUnsubscribe() {
+    return this.emailUnsubscribeDomain;
+  }
+
+  get health() {
+    return this.healthDomain;
+  }
+
+  get xRead() {
+    return this.xReadDomain;
+  }
+
+  get imessage() {
+    return this.imessageDomain;
+  }
+
+  get telegram() {
+    return this.telegramDomain;
+  }
+
+  get discord() {
+    return this.discordDomain;
+  }
+
+  get signal() {
+    return this.signalDomain;
+  }
+
+  get whatsapp() {
+    return this.whatsappDomain;
+  }
+
+  get travel() {
+    return this.travelDomain;
+  }
+
+  get scheduling() {
+    return this.schedulingDomain;
+  }
+
+  get subscriptions() {
+    return this.subscriptionsDomain;
+  }
+
+  get status() {
+    return this.statusDomain;
+  }
+
+  get screenTime() {
+    return this.screenTimeDomain;
+  }
+
+  get sleep() {
+    return this.sleepDomain;
+  }
+
+  get inbox() {
+    return this.inboxDomain;
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly googleDomain = new GoogleDomain(this);
+
+  public withGoogleGrantOperation<T>(
+    _grant: LifeOpsConnectorGrant,
+    operation: () => Promise<T>,
+  ): Promise<T> {
+    return this.googleDomain.withGoogleGrantOperation(_grant, operation);
+  }
+
+  public runManagedGoogleOperation<T>(
+    _grant: LifeOpsConnectorGrant,
+    _operation: () => Promise<T>,
+  ): Promise<T> {
+    return this.googleDomain.runManagedGoogleOperation(_grant, _operation);
+  }
+
+  public clearGoogleConnectorData(side?: LifeOpsConnectorSide): Promise<void> {
+    return this.googleDomain.clearGoogleConnectorData(side);
+  }
+
+  public clearGoogleGrantData(grant: LifeOpsConnectorGrant): Promise<void> {
+    return this.googleDomain.clearGoogleGrantData(grant);
+  }
+
+  public deleteCalendarReminderPlansForEvents(
+    _eventIds: string[],
+  ): Promise<void> {
+    return this.googleDomain.deleteCalendarReminderPlansForEvents(_eventIds);
+  }
+
+  public setPreferredGoogleConnectorMode(
+    _mode: LifeOpsConnectorMode | null,
+    _side?: LifeOpsConnectorSide,
+  ): Promise<LifeOpsConnectorGrant | null> {
+    return this.googleDomain.setPreferredGoogleConnectorMode(_mode, _side);
+  }
+
+  public requireGoogleCalendarGrant(
+    requestUrl: URL,
+    requestedMode?: LifeOpsConnectorMode,
+    requestedSide?: LifeOpsConnectorSide,
+    grantId?: string,
+  ): Promise<LifeOpsConnectorGrant> {
+    return this.googleDomain.requireGoogleCalendarGrant(
+      requestUrl,
+      requestedMode,
+      requestedSide,
+      grantId,
+    );
+  }
+
+  public requireGoogleCalendarWriteGrant(
+    requestUrl: URL,
+    requestedMode?: LifeOpsConnectorMode,
+    requestedSide?: LifeOpsConnectorSide,
+    grantId?: string,
+  ): Promise<LifeOpsConnectorGrant> {
+    return this.googleDomain.requireGoogleCalendarWriteGrant(
+      requestUrl,
+      requestedMode,
+      requestedSide,
+      grantId,
+    );
+  }
+
+  public requireGoogleGmailGrant(
+    requestUrl: URL,
+    requestedMode?: LifeOpsConnectorMode,
+    requestedSide?: LifeOpsConnectorSide,
+    grantId?: string,
+  ): Promise<LifeOpsConnectorGrant> {
+    return this.googleDomain.requireGoogleGmailGrant(
+      requestUrl,
+      requestedMode,
+      requestedSide,
+      grantId,
+    );
+  }
+
+  public requireGoogleGmailSendGrant(
+    requestUrl: URL,
+    requestedMode?: LifeOpsConnectorMode,
+    requestedSide?: LifeOpsConnectorSide,
+    grantId?: string,
+  ): Promise<LifeOpsConnectorGrant> {
+    return this.googleDomain.requireGoogleGmailSendGrant(
+      requestUrl,
+      requestedMode,
+      requestedSide,
+      grantId,
+    );
+  }
+
+  getGoogleConnectorStatus(
+    requestUrl: URL,
+    requestedMode?: LifeOpsConnectorMode,
+    requestedSide?: LifeOpsConnectorSide,
+    grantId?: string,
+  ): Promise<LifeOpsGoogleConnectorStatus> {
+    return this.googleDomain.getGoogleConnectorStatus(
+      requestUrl,
+      requestedMode,
+      requestedSide,
+      grantId,
+    );
+  }
+
+  getGoogleConnectorAccounts(
+    requestUrl: URL,
+    requestedSide?: LifeOpsConnectorSide,
+  ): Promise<LifeOpsGoogleConnectorStatus[]> {
+    return this.googleDomain.getGoogleConnectorAccounts(
+      requestUrl,
+      requestedSide,
+    );
+  }
+
+  selectGoogleConnectorMode(
+    requestUrl: URL,
+    preferredModeInput: LifeOpsConnectorMode | undefined,
+    requestedSide?: LifeOpsConnectorSide,
+  ): Promise<LifeOpsGoogleConnectorStatus> {
+    return this.googleDomain.selectGoogleConnectorMode(
+      requestUrl,
+      preferredModeInput,
+      requestedSide,
+    );
+  }
+
+  startGoogleConnector(
+    request: StartLifeOpsGoogleConnectorRequest,
+    requestUrl: URL,
+  ): Promise<StartLifeOpsGoogleConnectorResponse> {
+    return this.googleDomain.startGoogleConnector(request, requestUrl);
+  }
+
+  completeGoogleConnectorCallback(
+    callbackUrl: URL,
+  ): Promise<LifeOpsGoogleConnectorStatus> {
+    return this.googleDomain.completeGoogleConnectorCallback(callbackUrl);
+  }
+
+  disconnectGoogleConnector(
+    request: DisconnectLifeOpsGoogleConnectorRequest,
+    requestUrl: URL,
+  ): Promise<LifeOpsGoogleConnectorStatus> {
+    return this.googleDomain.disconnectGoogleConnector(request, requestUrl);
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly calendarDomain = new CalendarDomain(this);
+
+  listCalendars(
+    requestUrl: URL,
+    request?: ListLifeOpsCalendarsRequest,
+  ): Promise<LifeOpsCalendarSummary[]> {
+    return this.calendarDomain.listCalendars(requestUrl, request);
+  }
+
+  setCalendarIncluded(
+    requestUrl: URL,
+    request: {
+      calendarId: string;
+      includeInFeed: boolean;
+      side?: LifeOpsConnectorSide;
+      mode?: LifeOpsConnectorMode;
+      grantId?: string;
+    },
+  ): Promise<LifeOpsCalendarSummary> {
+    return this.calendarDomain.setCalendarIncluded(requestUrl, request);
+  }
+
+  getCalendarFeed(
+    requestUrl: URL,
+    request?: GetLifeOpsCalendarFeedRequest,
+    now?: Date,
+  ): Promise<LifeOpsCalendarFeed> {
+    return this.calendarDomain.getCalendarFeed(requestUrl, request, now);
+  }
+
+  createCalendarEvent(
+    requestUrl: URL,
+    request: CreateLifeOpsCalendarEventRequest,
+    now?: Date,
+  ): Promise<LifeOpsCalendarEvent> {
+    return this.calendarDomain.createCalendarEvent(requestUrl, request, now);
+  }
+
+  updateCalendarEvent(
+    requestUrl: URL,
+    request: {
+      mode?: LifeOpsConnectorMode | null;
+      side?: LifeOpsConnectorSide | null;
+      grantId?: string;
+      calendarId?: string | null;
+      eventId: string;
+      title?: string;
+      description?: string;
+      location?: string;
+      startAt?: string;
+      endAt?: string;
+      timeZone?: string;
+      attendees?: CreateLifeOpsCalendarEventAttendee[] | null;
+    },
+  ): Promise<LifeOpsCalendarEvent> {
+    return this.calendarDomain.updateCalendarEvent(requestUrl, request);
+  }
+
+  deleteCalendarEvent(
+    requestUrl: URL,
+    request: {
+      mode?: LifeOpsConnectorMode | null;
+      side?: LifeOpsConnectorSide | null;
+      grantId?: string;
+      calendarId?: string | null;
+      eventId: string;
+    },
+  ): Promise<void> {
+    return this.calendarDomain.deleteCalendarEvent(requestUrl, request);
+  }
+
+  getNextCalendarEventContext(
+    requestUrl: URL,
+    request?: GetLifeOpsCalendarFeedRequest,
+    now?: Date,
+  ): Promise<LifeOpsNextCalendarEventContext> {
+    return this.calendarDomain.getNextCalendarEventContext(
+      requestUrl,
+      request,
+      now,
+    );
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly gmailDomain = new GmailDomain(this, {
+    requireGoogleGmailGrant: (
+      requestUrl,
+      requestedMode,
+      requestedSide,
+      grantId,
+    ) =>
+      this.requireGoogleGmailGrant(
+        requestUrl,
+        requestedMode,
+        requestedSide,
+        grantId,
+      ),
+    requireGoogleGmailSendGrant: (
+      requestUrl,
+      requestedMode,
+      requestedSide,
+      grantId,
+    ) =>
+      this.requireGoogleGmailSendGrant(
+        requestUrl,
+        requestedMode,
+        requestedSide,
+        grantId,
+      ),
+  });
+
+  getGmailTriage(
+    requestUrl: URL,
+    request?: GetLifeOpsGmailTriageRequest,
+    now?: Date,
+  ): Promise<LifeOpsGmailTriageFeed> {
+    return this.gmailDomain.getGmailTriage(requestUrl, request, now);
+  }
+
+  getGmailSearch(
+    requestUrl: URL,
+    request: GetLifeOpsGmailSearchRequest,
+    now?: Date,
+  ): Promise<LifeOpsGmailSearchFeed> {
+    return this.gmailDomain.getGmailSearch(requestUrl, request, now);
+  }
+
+  readGmailMessage(
+    requestUrl: URL,
+    request: {
+      side?: LifeOpsConnectorSide;
+      mode?: LifeOpsConnectorMode;
+      grantId?: string;
+      forceSync?: boolean;
+      maxResults?: number;
+      messageId?: string;
+      query?: string;
+      replyNeededOnly?: boolean;
+    },
+    now?: Date,
+  ): Promise<{
+    query: string | null;
+    message: LifeOpsGmailMessageSummary;
+    bodyText: string;
+    source: "synced";
+    syncedAt: string;
+  }> {
+    return this.gmailDomain.readGmailMessage(requestUrl, request, now);
+  }
+
+  getGmailNeedsResponse(
+    requestUrl: URL,
+    request?: GetLifeOpsGmailTriageRequest,
+    now?: Date,
+  ): Promise<LifeOpsGmailNeedsResponseFeed> {
+    return this.gmailDomain.getGmailNeedsResponse(requestUrl, request, now);
+  }
+
+  getGmailRecommendations(
+    requestUrl: URL,
+    request?: GetLifeOpsGmailRecommendationsRequest,
+    now?: Date,
+  ): Promise<LifeOpsGmailRecommendationsFeed> {
+    return this.gmailDomain.getGmailRecommendations(requestUrl, request, now);
+  }
+
+  getGmailSpamReviewItems(
+    requestUrl: URL,
+    request?: GetLifeOpsGmailSpamReviewRequest,
+  ): Promise<LifeOpsGmailSpamReviewFeed> {
+    return this.gmailDomain.getGmailSpamReviewItems(requestUrl, request);
+  }
+
+  updateGmailSpamReviewItem(
+    requestUrl: URL,
+    itemId: string,
+    request: UpdateLifeOpsGmailSpamReviewItemRequest,
+    now?: Date,
+  ): Promise<{ item: LifeOpsGmailSpamReviewItem }> {
+    return this.gmailDomain.updateGmailSpamReviewItem(
+      requestUrl,
+      itemId,
+      request,
+      now,
+    );
+  }
+
+  getGmailUnresponded(
+    requestUrl: URL,
+    request?: GetLifeOpsGmailUnrespondedRequest,
+    now?: Date,
+  ): Promise<LifeOpsGmailUnrespondedFeed> {
+    return this.gmailDomain.getGmailUnresponded(requestUrl, request, now);
+  }
+
+  manageGmailMessages(
+    requestUrl: URL,
+    request: ManageLifeOpsGmailMessagesRequest,
+  ): Promise<LifeOpsGmailManageResult> {
+    return this.gmailDomain.manageGmailMessages(requestUrl, request);
+  }
+
+  ingestGmailEvent(
+    requestUrl: URL,
+    request: IngestLifeOpsGmailEventRequest,
+    now?: Date,
+  ): Promise<LifeOpsGmailEventIngestResult> {
+    return this.gmailDomain.ingestGmailEvent(requestUrl, request, now);
+  }
+
+  createGmailBatchReplyDrafts(
+    requestUrl: URL,
+    request: CreateLifeOpsGmailBatchReplyDraftsRequest,
+    now?: Date,
+  ): Promise<LifeOpsGmailBatchReplyDraftsFeed> {
+    return this.gmailDomain.createGmailBatchReplyDrafts(
+      requestUrl,
+      request,
+      now,
+    );
+  }
+
+  createGmailReplyDraft(
+    requestUrl: URL,
+    request: CreateLifeOpsGmailReplyDraftRequest,
+  ): Promise<LifeOpsGmailReplyDraft> {
+    return this.gmailDomain.createGmailReplyDraft(requestUrl, request);
+  }
+
+  sendGmailReply(
+    requestUrl: URL,
+    request: SendLifeOpsGmailReplyRequest,
+  ): Promise<{ ok: true }> {
+    return this.gmailDomain.sendGmailReply(requestUrl, request);
+  }
+
+  sendGmailMessage(
+    requestUrl: URL,
+    request: SendLifeOpsGmailMessageRequest,
+  ): Promise<{ ok: true }> {
+    return this.gmailDomain.sendGmailMessage(requestUrl, request);
+  }
+
+  sendGmailReplies(
+    requestUrl: URL,
+    request: SendLifeOpsGmailBatchReplyRequest,
+  ): Promise<LifeOpsGmailBatchReplySendResult> {
+    return this.gmailDomain.sendGmailReplies(requestUrl, request);
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly driveDomain = new DriveDomain(this, {
+    getGoogleConnectorStatus: (
+      requestUrl,
+      requestedMode,
+      requestedSide,
+      grantId,
+    ) =>
+      this.getGoogleConnectorStatus(
+        requestUrl,
+        requestedMode,
+        requestedSide,
+        grantId,
+      ),
+  });
+
+  requireGoogleDriveReadGrant(
+    requestUrl: URL,
+    requestedMode?: LifeOpsConnectorMode,
+    requestedSide?: LifeOpsConnectorSide,
+    grantId?: string,
+  ): Promise<LifeOpsConnectorGrant> {
+    return this.driveDomain.requireGoogleDriveReadGrant(
+      requestUrl,
+      requestedMode,
+      requestedSide,
+      grantId,
+    );
+  }
+
+  requireGoogleDriveWriteGrant(
+    requestUrl: URL,
+    requestedMode?: LifeOpsConnectorMode,
+    requestedSide?: LifeOpsConnectorSide,
+    grantId?: string,
+  ): Promise<LifeOpsConnectorGrant> {
+    return this.driveDomain.requireGoogleDriveWriteGrant(
+      requestUrl,
+      requestedMode,
+      requestedSide,
+      grantId,
+    );
+  }
+
+  listDriveFiles(
+    requestUrl: URL,
+    request?: {
+      mode?: LifeOpsConnectorMode;
+      side?: LifeOpsConnectorSide;
+      grantId?: string;
+      folderId?: string;
+      maxResults?: number;
+      pageToken?: string;
+    },
+  ): Promise<{ files: GoogleDriveFile[]; nextPageToken: string | null }> {
+    return this.driveDomain.listDriveFiles(requestUrl, request);
+  }
+
+  getDriveFile(
+    requestUrl: URL,
+    request: {
+      mode?: LifeOpsConnectorMode;
+      side?: LifeOpsConnectorSide;
+      grantId?: string;
+      fileId: string;
+    },
+  ): Promise<GoogleDriveFile> {
+    return this.driveDomain.getDriveFile(requestUrl, request);
+  }
+
+  searchDriveFiles(
+    requestUrl: URL,
+    request: {
+      mode?: LifeOpsConnectorMode;
+      side?: LifeOpsConnectorSide;
+      grantId?: string;
+      query: string;
+      maxResults?: number;
+    },
+  ): Promise<{ files: GoogleDriveFile[]; nextPageToken: string | null }> {
+    return this.driveDomain.searchDriveFiles(requestUrl, request);
+  }
+
+  getDocContent(
+    requestUrl: URL,
+    request: {
+      mode?: LifeOpsConnectorMode;
+      side?: LifeOpsConnectorSide;
+      grantId?: string;
+      documentId: string;
+    },
+  ): Promise<{ title: string; plainText: string }> {
+    return this.driveDomain.getDocContent(requestUrl, request);
+  }
+
+  getSheetContent(
+    requestUrl: URL,
+    request: {
+      mode?: LifeOpsConnectorMode;
+      side?: LifeOpsConnectorSide;
+      grantId?: string;
+      spreadsheetId: string;
+      range?: string;
+    },
+  ): Promise<{ title: string; rows: string[][] }> {
+    return this.driveDomain.getSheetContent(requestUrl, request);
+  }
+
+  createDriveFile(
+    requestUrl: URL,
+    request: {
+      mode?: LifeOpsConnectorMode;
+      side?: LifeOpsConnectorSide;
+      grantId?: string;
+      name: string;
+      mimeType: string;
+      content?: string | Uint8Array;
+      parentFolderId?: string;
+    },
+  ): Promise<GoogleDriveFile> {
+    return this.driveDomain.createDriveFile(requestUrl, request);
+  }
+
+  appendToDoc(
+    requestUrl: URL,
+    request: {
+      mode?: LifeOpsConnectorMode;
+      side?: LifeOpsConnectorSide;
+      grantId?: string;
+      documentId: string;
+      text: string;
+    },
+  ): Promise<void> {
+    return this.driveDomain.appendToDoc(requestUrl, request);
+  }
+
+  updateSheetCells(
+    requestUrl: URL,
+    request: {
+      mode?: LifeOpsConnectorMode;
+      side?: LifeOpsConnectorSide;
+      grantId?: string;
+      spreadsheetId: string;
+      range: string;
+      values: ReadonlyArray<ReadonlyArray<string | number>>;
+    },
+  ): Promise<{ updatedRange: string; updatedCells: number }> {
+    return this.driveDomain.updateSheetCells(requestUrl, request);
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext, and the
+  // composed runtime service supplies the cross-domain dependencies below.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly remindersDomain = new RemindersDomain(this, {
+    runDueWorkflows: (...args) => this.runDueWorkflows(...args),
+    // RemindersDeps types the event payload as the health-domain
+    // `LifeOpsDerivedEvent`, while the workflows domain method this forwards to
+    // types it as `LifeOpsWorkflowEvent`. The two are runtime-compatible; a
+    // localized structural cast reconciles the cross-domain payload types.
+    runDueEventWorkflows: (...args) =>
+      (
+        this as unknown as {
+          runDueEventWorkflows(args: {
+            now: string;
+            limit: number;
+            lifeOpsEvents?: LifeOpsDerivedEvent[];
+          }): Promise<LifeOpsWorkflowRun[]>;
+        }
+      ).runDueEventWorkflows(...args),
+    snoozeOccurrence: (...args) => this.snoozeOccurrence(...args),
+    checkinSource: this,
+  });
+
+  readRecentReminderConversation(
+    ...args: Parameters<RemindersDomain["readRecentReminderConversation"]>
+  ): ReturnType<RemindersDomain["readRecentReminderConversation"]> {
+    return this.remindersDomain.readRecentReminderConversation(...args);
+  }
+
+  classifyReminderOwnerResponseSemantically(
+    ...args: Parameters<
+      RemindersDomain["classifyReminderOwnerResponseSemantically"]
+    >
+  ): ReturnType<RemindersDomain["classifyReminderOwnerResponseSemantically"]> {
+    return this.remindersDomain.classifyReminderOwnerResponseSemantically(
+      ...args,
+    );
+  }
+
+  reviewOwnerResponseAfterReminderAttempt(
+    ...args: Parameters<
+      RemindersDomain["reviewOwnerResponseAfterReminderAttempt"]
+    >
+  ): ReturnType<RemindersDomain["reviewOwnerResponseAfterReminderAttempt"]> {
+    return this.remindersDomain.reviewOwnerResponseAfterReminderAttempt(
+      ...args,
+    );
+  }
+
+  renderReminderBody(
+    ...args: Parameters<RemindersDomain["renderReminderBody"]>
+  ): ReturnType<RemindersDomain["renderReminderBody"]> {
+    return this.remindersDomain.renderReminderBody(...args);
+  }
+
+  renderWorkflowRunBody(
+    ...args: Parameters<RemindersDomain["renderWorkflowRunBody"]>
+  ): ReturnType<RemindersDomain["renderWorkflowRunBody"]> {
+    return this.remindersDomain.renderWorkflowRunBody(...args);
+  }
+
+  emitWorkflowRunNudge(
+    ...args: Parameters<RemindersDomain["emitWorkflowRunNudge"]>
+  ): ReturnType<RemindersDomain["emitWorkflowRunNudge"]> {
+    return this.remindersDomain.emitWorkflowRunNudge(...args);
+  }
+
+  withNativeAppleReminderId(
+    ...args: Parameters<RemindersDomain["withNativeAppleReminderId"]>
+  ): ReturnType<RemindersDomain["withNativeAppleReminderId"]> {
+    return this.remindersDomain.withNativeAppleReminderId(...args);
+  }
+
+  syncNativeAppleReminderForDefinition(
+    ...args: Parameters<RemindersDomain["syncNativeAppleReminderForDefinition"]>
+  ): ReturnType<RemindersDomain["syncNativeAppleReminderForDefinition"]> {
+    return this.remindersDomain.syncNativeAppleReminderForDefinition(...args);
+  }
+
+  getDefinitionRecord(
+    ...args: Parameters<RemindersDomain["getDefinitionRecord"]>
+  ): ReturnType<RemindersDomain["getDefinitionRecord"]> {
+    return this.remindersDomain.getDefinitionRecord(...args);
+  }
+
+  getGoalRecord(
+    ...args: Parameters<RemindersDomain["getGoalRecord"]>
+  ): ReturnType<RemindersDomain["getGoalRecord"]> {
+    return this.remindersDomain.getGoalRecord(...args);
+  }
+
+  ensureGoalExists(
+    ...args: Parameters<RemindersDomain["ensureGoalExists"]>
+  ): ReturnType<RemindersDomain["ensureGoalExists"]> {
+    return this.remindersDomain.ensureGoalExists(...args);
+  }
+
+  syncGoalLink(
+    ...args: Parameters<RemindersDomain["syncGoalLink"]>
+  ): ReturnType<RemindersDomain["syncGoalLink"]> {
+    return this.remindersDomain.syncGoalLink(...args);
+  }
+
+  syncReminderPlan(
+    ...args: Parameters<RemindersDomain["syncReminderPlan"]>
+  ): ReturnType<RemindersDomain["syncReminderPlan"]> {
+    return this.remindersDomain.syncReminderPlan(...args);
+  }
+
+  serializeScheduleObservationForSync(
+    ...args: Parameters<RemindersDomain["serializeScheduleObservationForSync"]>
+  ): ReturnType<RemindersDomain["serializeScheduleObservationForSync"]> {
+    return this.remindersDomain.serializeScheduleObservationForSync(...args);
+  }
+
+  refreshLocalMergedScheduleState(
+    ...args: Parameters<RemindersDomain["refreshLocalMergedScheduleState"]>
+  ): ReturnType<RemindersDomain["refreshLocalMergedScheduleState"]> {
+    return this.remindersDomain.refreshLocalMergedScheduleState(...args);
+  }
+
+  ingestScheduleObservations(
+    ...args: Parameters<RemindersDomain["ingestScheduleObservations"]>
+  ): ReturnType<RemindersDomain["ingestScheduleObservations"]> {
+    return this.remindersDomain.ingestScheduleObservations(...args);
+  }
+
+  fetchCloudMergedScheduleState(
+    ...args: Parameters<RemindersDomain["fetchCloudMergedScheduleState"]>
+  ): ReturnType<RemindersDomain["fetchCloudMergedScheduleState"]> {
+    return this.remindersDomain.fetchCloudMergedScheduleState(...args);
+  }
+
+  readEffectiveScheduleState(
+    ...args: Parameters<RemindersDomain["readEffectiveScheduleState"]>
+  ): ReturnType<RemindersDomain["readEffectiveScheduleState"]> {
+    return this.remindersDomain.readEffectiveScheduleState(...args);
+  }
+
+  refreshEffectiveScheduleState(
+    ...args: Parameters<RemindersDomain["refreshEffectiveScheduleState"]>
+  ): ReturnType<RemindersDomain["refreshEffectiveScheduleState"]> {
+    return this.remindersDomain.refreshEffectiveScheduleState(...args);
+  }
+
+  getScheduleMergedState(
+    ...args: Parameters<RemindersDomain["getScheduleMergedState"]>
+  ): ReturnType<RemindersDomain["getScheduleMergedState"]> {
+    return this.remindersDomain.getScheduleMergedState(...args);
+  }
+
+  resolveAdaptiveWindowPolicy(
+    ...args: Parameters<RemindersDomain["resolveAdaptiveWindowPolicy"]>
+  ): ReturnType<RemindersDomain["resolveAdaptiveWindowPolicy"]> {
+    return this.remindersDomain.resolveAdaptiveWindowPolicy(...args);
+  }
+
+  refreshDefinitionOccurrences(
+    ...args: Parameters<RemindersDomain["refreshDefinitionOccurrences"]>
+  ): ReturnType<RemindersDomain["refreshDefinitionOccurrences"]> {
+    return this.remindersDomain.refreshDefinitionOccurrences(...args);
+  }
+
+  getFreshOccurrence(
+    ...args: Parameters<RemindersDomain["getFreshOccurrence"]>
+  ): ReturnType<RemindersDomain["getFreshOccurrence"]> {
+    return this.remindersDomain.getFreshOccurrence(...args);
+  }
+
+  resolvePrimaryChannelPolicy(
+    ...args: Parameters<RemindersDomain["resolvePrimaryChannelPolicy"]>
+  ): ReturnType<RemindersDomain["resolvePrimaryChannelPolicy"]> {
+    return this.remindersDomain.resolvePrimaryChannelPolicy(...args);
+  }
+
+  resolveRuntimeReminderTarget(
+    ...args: Parameters<RemindersDomain["resolveRuntimeReminderTarget"]>
+  ): ReturnType<RemindersDomain["resolveRuntimeReminderTarget"]> {
+    return this.remindersDomain.resolveRuntimeReminderTarget(...args);
+  }
+
+  readLifeOpsAttentionContext(
+    ...args: Parameters<RemindersDomain["readLifeOpsAttentionContext"]>
+  ): ReturnType<RemindersDomain["readLifeOpsAttentionContext"]> {
+    return this.remindersDomain.readLifeOpsAttentionContext(...args);
+  }
+
+  readReminderActivityProfileSnapshot(
+    ...args: Parameters<RemindersDomain["readReminderActivityProfileSnapshot"]>
+  ): ReturnType<RemindersDomain["readReminderActivityProfileSnapshot"]> {
+    return this.remindersDomain.readReminderActivityProfileSnapshot(...args);
+  }
+
+  scanReadReceipts(
+    ...args: Parameters<RemindersDomain["scanReadReceipts"]>
+  ): ReturnType<RemindersDomain["scanReadReceipts"]> {
+    return this.remindersDomain.scanReadReceipts(...args);
+  }
+
+  buildReminderPlanSchedule(
+    ...args: Parameters<RemindersDomain["buildReminderPlanSchedule"]>
+  ): ReturnType<RemindersDomain["buildReminderPlanSchedule"]> {
+    return this.remindersDomain.buildReminderPlanSchedule(...args);
+  }
+
+  resolveOwnerContactRouteCandidates(
+    ...args: Parameters<RemindersDomain["resolveOwnerContactRouteCandidates"]>
+  ): ReturnType<RemindersDomain["resolveOwnerContactRouteCandidates"]> {
+    return this.remindersDomain.resolveOwnerContactRouteCandidates(...args);
+  }
+
+  resolveReminderEscalationRouteCandidates(
+    ...args: Parameters<
+      RemindersDomain["resolveReminderEscalationRouteCandidates"]
+    >
+  ): ReturnType<RemindersDomain["resolveReminderEscalationRouteCandidates"]> {
+    return this.remindersDomain.resolveReminderEscalationRouteCandidates(
+      ...args,
+    );
+  }
+
+  buildOwnerContactRouteEventMetadata(
+    ...args: Parameters<RemindersDomain["buildOwnerContactRouteEventMetadata"]>
+  ): ReturnType<RemindersDomain["buildOwnerContactRouteEventMetadata"]> {
+    return this.remindersDomain.buildOwnerContactRouteEventMetadata(...args);
+  }
+
+  resolveReminderEscalationChannels(
+    ...args: Parameters<RemindersDomain["resolveReminderEscalationChannels"]>
+  ): ReturnType<RemindersDomain["resolveReminderEscalationChannels"]> {
+    return this.remindersDomain.resolveReminderEscalationChannels(...args);
+  }
+
+  markReminderEscalationStarted(
+    ...args: Parameters<RemindersDomain["markReminderEscalationStarted"]>
+  ): ReturnType<RemindersDomain["markReminderEscalationStarted"]> {
+    return this.remindersDomain.markReminderEscalationStarted(...args);
+  }
+
+  resolveReminderEscalation(
+    ...args: Parameters<RemindersDomain["resolveReminderEscalation"]>
+  ): ReturnType<RemindersDomain["resolveReminderEscalation"]> {
+    return this.remindersDomain.resolveReminderEscalation(...args);
+  }
+
+  resolveReminderReviewFromOwnerResponse(
+    ...args: Parameters<
+      RemindersDomain["resolveReminderReviewFromOwnerResponse"]
+    >
+  ): ReturnType<RemindersDomain["resolveReminderReviewFromOwnerResponse"]> {
+    return this.remindersDomain.resolveReminderReviewFromOwnerResponse(...args);
+  }
+
+  markReminderReviewResolvedFromState(
+    ...args: Parameters<RemindersDomain["markReminderReviewResolvedFromState"]>
+  ): ReturnType<RemindersDomain["markReminderReviewResolvedFromState"]> {
+    return this.remindersDomain.markReminderReviewResolvedFromState(...args);
+  }
+
+  markReminderReviewEscalated(
+    ...args: Parameters<RemindersDomain["markReminderReviewEscalated"]>
+  ): ReturnType<RemindersDomain["markReminderReviewEscalated"]> {
+    return this.remindersDomain.markReminderReviewEscalated(...args);
+  }
+
+  markReminderReviewClarificationRequested(
+    ...args: Parameters<
+      RemindersDomain["markReminderReviewClarificationRequested"]
+    >
+  ): ReturnType<RemindersDomain["markReminderReviewClarificationRequested"]> {
+    return this.remindersDomain.markReminderReviewClarificationRequested(
+      ...args,
+    );
+  }
+
+  markReminderReviewObservedResponse(
+    ...args: Parameters<RemindersDomain["markReminderReviewObservedResponse"]>
+  ): ReturnType<RemindersDomain["markReminderReviewObservedResponse"]> {
+    return this.remindersDomain.markReminderReviewObservedResponse(...args);
+  }
+
+  processDueReminderReviewJobs(
+    ...args: Parameters<RemindersDomain["processDueReminderReviewJobs"]>
+  ): ReturnType<RemindersDomain["processDueReminderReviewJobs"]> {
+    return this.remindersDomain.processDueReminderReviewJobs(...args);
+  }
+
+  dispatchDueReminderEscalation(
+    ...args: Parameters<RemindersDomain["dispatchDueReminderEscalation"]>
+  ): ReturnType<RemindersDomain["dispatchDueReminderEscalation"]> {
+    return this.remindersDomain.dispatchDueReminderEscalation(...args);
+  }
+
+  awardWebsiteAccessGrant(
+    ...args: Parameters<RemindersDomain["awardWebsiteAccessGrant"]>
+  ): ReturnType<RemindersDomain["awardWebsiteAccessGrant"]> {
+    return this.remindersDomain.awardWebsiteAccessGrant(...args);
+  }
+
+  syncWebsiteAccessState(
+    ...args: Parameters<RemindersDomain["syncWebsiteAccessState"]>
+  ): ReturnType<RemindersDomain["syncWebsiteAccessState"]> {
+    return this.remindersDomain.syncWebsiteAccessState(...args);
+  }
+
+  dispatchReminderAttempt(
+    ...args: Parameters<RemindersDomain["dispatchReminderAttempt"]>
+  ): ReturnType<RemindersDomain["dispatchReminderAttempt"]> {
+    return this.remindersDomain.dispatchReminderAttempt(...args);
+  }
+
+  resolveGlobalReminderPreferencePolicy(
+    ...args: Parameters<
+      RemindersDomain["resolveGlobalReminderPreferencePolicy"]
+    >
+  ): ReturnType<RemindersDomain["resolveGlobalReminderPreferencePolicy"]> {
+    return this.remindersDomain.resolveGlobalReminderPreferencePolicy(...args);
+  }
+
+  buildReminderPreferenceResponse(
+    ...args: Parameters<RemindersDomain["buildReminderPreferenceResponse"]>
+  ): ReturnType<RemindersDomain["buildReminderPreferenceResponse"]> {
+    return this.remindersDomain.buildReminderPreferenceResponse(...args);
+  }
+
+  resolveEffectiveReminderPlan(
+    ...args: Parameters<RemindersDomain["resolveEffectiveReminderPlan"]>
+  ): ReturnType<RemindersDomain["resolveEffectiveReminderPlan"]> {
+    return this.remindersDomain.resolveEffectiveReminderPlan(...args);
+  }
+
+  getReminderPreference(
+    ...args: Parameters<RemindersDomain["getReminderPreference"]>
+  ): ReturnType<RemindersDomain["getReminderPreference"]> {
+    return this.remindersDomain.getReminderPreference(...args);
+  }
+
+  setReminderPreference(
+    ...args: Parameters<RemindersDomain["setReminderPreference"]>
+  ): ReturnType<RemindersDomain["setReminderPreference"]> {
+    return this.remindersDomain.setReminderPreference(...args);
+  }
+
+  captureActivitySignal(
+    ...args: Parameters<RemindersDomain["captureActivitySignal"]>
+  ): ReturnType<RemindersDomain["captureActivitySignal"]> {
+    return this.remindersDomain.captureActivitySignal(...args);
+  }
+
+  captureManualOverride(
+    ...args: Parameters<RemindersDomain["captureManualOverride"]>
+  ): ReturnType<RemindersDomain["captureManualOverride"]> {
+    return this.remindersDomain.captureManualOverride(...args);
+  }
+
+  listActivitySignals(
+    ...args: Parameters<RemindersDomain["listActivitySignals"]>
+  ): ReturnType<RemindersDomain["listActivitySignals"]> {
+    return this.remindersDomain.listActivitySignals(...args);
+  }
+
+  upsertChannelPolicy(
+    ...args: Parameters<RemindersDomain["upsertChannelPolicy"]>
+  ): ReturnType<RemindersDomain["upsertChannelPolicy"]> {
+    return this.remindersDomain.upsertChannelPolicy(...args);
+  }
+
+  capturePhoneConsent(
+    ...args: Parameters<RemindersDomain["capturePhoneConsent"]>
+  ): ReturnType<RemindersDomain["capturePhoneConsent"]> {
+    return this.remindersDomain.capturePhoneConsent(...args);
+  }
+
+  processDueReminderDeliveries(
+    ...args: Parameters<RemindersDomain["processDueReminderDeliveries"]>
+  ): ReturnType<RemindersDomain["processDueReminderDeliveries"]> {
+    return this.remindersDomain.processDueReminderDeliveries(...args);
+  }
+
+  processReminders(
+    ...args: Parameters<RemindersDomain["processReminders"]>
+  ): ReturnType<RemindersDomain["processReminders"]> {
+    return this.remindersDomain.processReminders(...args);
+  }
+
+  processScheduledWork(
+    ...args: Parameters<RemindersDomain["processScheduledWork"]>
+  ): ReturnType<RemindersDomain["processScheduledWork"]> {
+    return this.remindersDomain.processScheduledWork(...args);
+  }
+
+  relockWebsiteAccessGroup(
+    ...args: Parameters<RemindersDomain["relockWebsiteAccessGroup"]>
+  ): ReturnType<RemindersDomain["relockWebsiteAccessGroup"]> {
+    return this.remindersDomain.relockWebsiteAccessGroup(...args);
+  }
+
+  resolveWebsiteAccessCallback(
+    ...args: Parameters<RemindersDomain["resolveWebsiteAccessCallback"]>
+  ): ReturnType<RemindersDomain["resolveWebsiteAccessCallback"]> {
+    return this.remindersDomain.resolveWebsiteAccessCallback(...args);
+  }
+
+  inspectReminder(
+    ...args: Parameters<RemindersDomain["inspectReminder"]>
+  ): ReturnType<RemindersDomain["inspectReminder"]> {
+    return this.remindersDomain.inspectReminder(...args);
+  }
+
+  acknowledgeReminder(
+    ...args: Parameters<RemindersDomain["acknowledgeReminder"]>
+  ): ReturnType<RemindersDomain["acknowledgeReminder"]> {
+    return this.remindersDomain.acknowledgeReminder(...args);
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly browserDomain = new BrowserDomain(this, {
+    getBrowserSettingsInternal: (...args) =>
+      this.getBrowserSettingsInternal(...args),
+    isBrowserPaused: (...args) => this.isBrowserPaused(...args),
+    requireBrowserAvailableForActions: (...args) =>
+      this.requireBrowserAvailableForActions(...args),
+    buildBrowserCompanion: (...args) => this.buildBrowserCompanion(...args),
+    recordBrowserAudit: (...args) => this.recordBrowserAudit(...args),
+    getWorkflowDefinition: (...args) => this.getWorkflowDefinition(...args),
+    recordScreenTimeEvent: (...args) => this.recordScreenTimeEvent(...args),
+  });
+
+  createBrowserSessionInternal(
+    request: CreateLifeOpsBrowserSessionRequest,
+  ): Promise<LifeOpsBrowserSession> {
+    return this.browserDomain.createBrowserSessionInternal(request);
+  }
+
+  getBrowserSettings(): Promise<BrowserBridgeSettings> {
+    return this.browserDomain.getBrowserSettings();
+  }
+
+  updateBrowserSettings(
+    request: UpdateBrowserBridgeSettingsRequest,
+  ): Promise<BrowserBridgeSettings> {
+    return this.browserDomain.updateBrowserSettings(request);
+  }
+
+  listBrowserCompanions(): Promise<BrowserBridgeCompanionStatus[]> {
+    return this.browserDomain.listBrowserCompanions();
+  }
+
+  listBrowserTabs(): Promise<BrowserBridgeTabSummary[]> {
+    return this.browserDomain.listBrowserTabs();
+  }
+
+  getCurrentBrowserPage(): Promise<BrowserBridgePageContext | null> {
+    return this.browserDomain.getCurrentBrowserPage();
+  }
+
+  syncBrowserState(request: SyncBrowserBridgeStateRequest): Promise<{
+    companion: BrowserBridgeCompanionStatus;
+    tabs: BrowserBridgeTabSummary[];
+    currentPage: BrowserBridgePageContext | null;
+  }> {
+    return this.browserDomain.syncBrowserState(request);
+  }
+
+  createBrowserCompanionPairing(
+    request: CreateBrowserBridgeCompanionPairingRequest,
+  ): Promise<BrowserBridgeCompanionPairingResponse> {
+    return this.browserDomain.createBrowserCompanionPairing(request);
+  }
+
+  syncBrowserCompanion(
+    companionId: string,
+    pairingToken: string,
+    request: SyncBrowserBridgeStateRequest,
+  ): Promise<BrowserBridgeCompanionSyncResponse> {
+    return this.browserDomain.syncBrowserCompanion(
+      companionId,
+      pairingToken,
+      request,
+    );
+  }
+
+  listBrowserSessions(): Promise<LifeOpsBrowserSession[]> {
+    return this.browserDomain.listBrowserSessions();
+  }
+
+  getBrowserSession(sessionId: string): Promise<LifeOpsBrowserSession> {
+    return this.browserDomain.getBrowserSession(sessionId);
+  }
+
+  createBrowserSession(
+    request: CreateLifeOpsBrowserSessionRequest,
+  ): Promise<LifeOpsBrowserSession> {
+    return this.browserDomain.createBrowserSession(request);
+  }
+
+  confirmBrowserSession(
+    sessionId: string,
+    request: ConfirmLifeOpsBrowserSessionRequest,
+  ): Promise<LifeOpsBrowserSession> {
+    return this.browserDomain.confirmBrowserSession(sessionId, request);
+  }
+
+  completeBrowserSession(
+    sessionId: string,
+    request: CompleteLifeOpsBrowserSessionRequest,
+  ): Promise<LifeOpsBrowserSession> {
+    return this.browserDomain.completeBrowserSession(sessionId, request);
+  }
+
+  updateBrowserSessionProgress(
+    sessionId: string,
+    request: UpdateLifeOpsBrowserSessionProgressRequest,
+  ): Promise<LifeOpsBrowserSession> {
+    return this.browserDomain.updateBrowserSessionProgress(sessionId, request);
+  }
+
+  updateBrowserSessionProgressFromCompanion(
+    companionId: string,
+    pairingToken: string,
+    sessionId: string,
+    request: UpdateLifeOpsBrowserSessionProgressRequest,
+  ): Promise<LifeOpsBrowserSession> {
+    return this.browserDomain.updateBrowserSessionProgressFromCompanion(
+      companionId,
+      pairingToken,
+      sessionId,
+      request,
+    );
+  }
+
+  completeBrowserSessionFromCompanion(
+    companionId: string,
+    pairingToken: string,
+    sessionId: string,
+    request: CompleteLifeOpsBrowserSessionRequest,
+  ): Promise<LifeOpsBrowserSession> {
+    return this.browserDomain.completeBrowserSessionFromCompanion(
+      companionId,
+      pairingToken,
+      sessionId,
+      request,
+    );
+  }
+
+  autoPairBrowserCompanion(
+    request: CreateBrowserBridgeCompanionAutoPairRequest,
+    apiBaseUrl: string,
+  ): Promise<BrowserBridgeCompanionAutoPairResponse> {
+    return this.browserDomain.autoPairBrowserCompanion(request, apiBaseUrl);
+  }
+
+  revokeBrowserCompanion(
+    companionId: string,
+  ): Promise<BrowserBridgeCompanionRevokeResponse> {
+    return this.browserDomain.revokeBrowserCompanion(companionId);
+  }
+
+  revokeBrowserCompanionFromCompanion(
+    companionId: string,
+    pairingToken: string,
+  ): Promise<BrowserBridgeCompanionRevokeResponse> {
+    return this.browserDomain.revokeBrowserCompanionFromCompanion(
+      companionId,
+      pairingToken,
+    );
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly workflowsDomain = new WorkflowsDomain(this, {
+    recordWorkflowAudit: (...args) => this.recordWorkflowAudit(...args),
+    getWorkflowDefinition: (...args) => this.getWorkflowDefinition(...args),
+    readEffectiveScheduleState: (...args) =>
+      this.readEffectiveScheduleState(...args),
+    emitWorkflowRunNudge: (...args) => this.emitWorkflowRunNudge(...args),
+    // Workflow-step contributions reach across many domains, so the
+    // execution context is the fully composed service instance, not the
+    // workflows sub-service.
+    workflowStepContext: this,
+  });
+
+  listWorkflows(): Promise<LifeOpsWorkflowRecord[]> {
+    return this.workflowsDomain.listWorkflows();
+  }
+
+  getWorkflow(workflowId: string): Promise<LifeOpsWorkflowRecord> {
+    return this.workflowsDomain.getWorkflow(workflowId);
+  }
+
+  createWorkflow(
+    request: CreateLifeOpsWorkflowRequest,
+  ): Promise<LifeOpsWorkflowRecord> {
+    return this.workflowsDomain.createWorkflow(request);
+  }
+
+  updateWorkflow(
+    workflowId: string,
+    request: UpdateLifeOpsWorkflowRequest,
+  ): Promise<LifeOpsWorkflowRecord> {
+    return this.workflowsDomain.updateWorkflow(workflowId, request);
+  }
+
+  runWorkflow(
+    workflowId: string,
+    request: { now?: string; confirmBrowserActions?: boolean } = {},
+  ): Promise<LifeOpsWorkflowRun> {
+    return this.workflowsDomain.runWorkflow(workflowId, request);
+  }
+
+  // Consumed by the reminders scheduler via the composed instance.
+  runDueWorkflows(
+    args: Parameters<WorkflowsDomain["runDueWorkflows"]>[0],
+  ): Promise<LifeOpsWorkflowRun[]> {
+    return this.workflowsDomain.runDueWorkflows(args);
+  }
+
+  runDueEventWorkflows(
+    args: Parameters<WorkflowsDomain["runDueEventWorkflows"]>[0],
+  ): Promise<LifeOpsWorkflowRun[]> {
+    return this.workflowsDomain.runDueEventWorkflows(args);
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly definitionsDomain = new DefinitionsDomain(this, {
+    getDefinitionRecord: (...args) => this.getDefinitionRecord(...args),
+    ensureGoalExists: (...args) => this.ensureGoalExists(...args),
+    syncReminderPlan: (...args) => this.syncReminderPlan(...args),
+    syncGoalLink: (...args) => this.syncGoalLink(...args),
+    refreshDefinitionOccurrences: (...args) =>
+      this.refreshDefinitionOccurrences(...args),
+    syncNativeAppleReminderForDefinition: (...args) =>
+      this.syncNativeAppleReminderForDefinition(...args),
+    syncWebsiteAccessState: (...args) => this.syncWebsiteAccessState(...args),
+    getFreshOccurrence: (...args) => this.getFreshOccurrence(...args),
+    awardWebsiteAccessGrant: (...args) => this.awardWebsiteAccessGrant(...args),
+    resolveReminderEscalation: (...args) =>
+      this.resolveReminderEscalation(...args),
+  });
+
+  listDefinitions(): Promise<LifeOpsDefinitionRecord[]> {
+    return this.definitionsDomain.listDefinitions();
+  }
+
+  getDefinition(definitionId: string): Promise<LifeOpsDefinitionRecord> {
+    return this.definitionsDomain.getDefinition(definitionId);
+  }
+
+  createDefinition(
+    request: CreateLifeOpsDefinitionRequest,
+  ): Promise<LifeOpsDefinitionRecord> {
+    return this.definitionsDomain.createDefinition(request);
+  }
+
+  updateDefinition(
+    definitionId: string,
+    request: UpdateLifeOpsDefinitionRequest,
+  ): Promise<LifeOpsDefinitionRecord> {
+    return this.definitionsDomain.updateDefinition(definitionId, request);
+  }
+
+  deleteDefinition(definitionId: string): Promise<void> {
+    return this.definitionsDomain.deleteDefinition(definitionId);
+  }
+
+  completeOccurrence(
+    occurrenceId: string,
+    request: CompleteLifeOpsOccurrenceRequest,
+    now?: Date,
+  ): Promise<LifeOpsOccurrenceView> {
+    return this.definitionsDomain.completeOccurrence(
+      occurrenceId,
+      request,
+      now,
+    );
+  }
+
+  skipOccurrence(
+    occurrenceId: string,
+    now?: Date,
+  ): Promise<LifeOpsOccurrenceView> {
+    return this.definitionsDomain.skipOccurrence(occurrenceId, now);
+  }
+
+  snoozeOccurrence(
+    occurrenceId: string,
+    request: SnoozeLifeOpsOccurrenceRequest,
+    now?: Date,
+  ): Promise<LifeOpsOccurrenceView> {
+    return this.definitionsDomain.snoozeOccurrence(occurrenceId, request, now);
+  }
+
+  readonly goalsDomain = new GoalsDomain(this, {
+    // The reminders domain re-declares a structurally-narrower local
+    // `LifeOpsGoalRecord`; GoalsDeps expects the shared contracts record. A
+    // localized structural cast reconciles the two record shapes.
+    getGoalRecord: (...args) =>
+      (
+        this as unknown as {
+          getGoalRecord(goalId: string): Promise<LifeOpsGoalRecord>;
+        }
+      ).getGoalRecord(...args),
+    getDefinitionRecord: (...args) => this.getDefinitionRecord(...args),
+    listActivitySignals: (...args) => this.listActivitySignals(...args),
+    inspectReminder: (...args) => this.inspectReminder(...args),
+    refreshEffectiveScheduleState: (...args) =>
+      this.refreshEffectiveScheduleState(...args),
+    refreshDefinitionOccurrences: (...args) =>
+      this.refreshDefinitionOccurrences(...args),
+    buildReminderPreferenceResponse: (...args) =>
+      this.buildReminderPreferenceResponse(...args),
+    resolveEffectiveReminderPlan: (...args) =>
+      this.resolveEffectiveReminderPlan(...args),
+  });
+
+  async deleteGoal(goalId: string): Promise<void> {
+    return this.goalsDomain.deleteGoal(goalId);
+  }
+
+  async listGoals(): Promise<LifeOpsGoalRecord[]> {
+    return this.goalsDomain.listGoals();
+  }
+
+  async getGoal(goalId: string): Promise<LifeOpsGoalRecord> {
+    return this.goalsDomain.getGoal(goalId);
+  }
+
+  async createGoal(
+    request: CreateLifeOpsGoalRequest,
+  ): Promise<LifeOpsGoalRecord> {
+    return this.goalsDomain.createGoal(request);
+  }
+
+  async updateGoal(
+    goalId: string,
+    request: UpdateLifeOpsGoalRequest,
+  ): Promise<LifeOpsGoalRecord> {
+    return this.goalsDomain.updateGoal(goalId, request);
+  }
+
+  async reviewGoal(
+    goalId: string,
+    now = new Date(),
+  ): Promise<LifeOpsGoalReview> {
+    return this.goalsDomain.reviewGoal(goalId, now);
+  }
+
+  async buildGoalExperienceLoop(
+    reference: {
+      goalId?: string | null;
+      title: string;
+      description?: string | null;
+      successCriteria?: Record<string, unknown> | null;
+    },
+    now = new Date(),
+  ): Promise<LifeOpsGoalExperienceLoop> {
+    return this.goalsDomain.buildGoalExperienceLoop(reference, now);
+  }
+
+  async reviewGoalsForWeek(now = new Date()): Promise<LifeOpsWeeklyGoalReview> {
+    return this.goalsDomain.reviewGoalsForWeek(now);
+  }
+
+  async explainOccurrence(
+    occurrenceId: string,
+  ): Promise<LifeOpsOccurrenceExplanation> {
+    return this.goalsDomain.explainOccurrence(occurrenceId);
+  }
+
+  async getOverview(now = new Date()): Promise<LifeOpsOverview> {
+    return this.goalsDomain.getOverview(now);
+  }
+
+  async listChannelPolicies(): Promise<LifeOpsChannelPolicy[]> {
+    return this.goalsDomain.listChannelPolicies();
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly xDomain = new XDomain(this, {
+    recordXPostAudit: (...args) => this.recordXPostAudit(...args),
+    resolvePrimaryChannelPolicy: (...args) =>
+      this.resolvePrimaryChannelPolicy(...args),
+  });
+
+  resolveXGrant(
+    requestedMode?: LifeOpsConnectorMode,
+    requestedSide?: LifeOpsConnectorSide,
+    requestedAccountId?: string | null,
+  ): Promise<LifeOpsConnectorGrant | null> {
+    return this.xDomain.resolveXGrant(
+      requestedMode,
+      requestedSide,
+      requestedAccountId,
+    );
+  }
+
+  getXConnectorStatus(
+    requestedMode?: LifeOpsConnectorMode,
+    requestedSide?: LifeOpsConnectorSide,
+    requestedAccountId?: string | null,
+  ): Promise<LifeOpsXConnectorStatus> {
+    return this.xDomain.getXConnectorStatus(
+      requestedMode,
+      requestedSide,
+      requestedAccountId,
+    );
+  }
+
+  createXPost(
+    request: CreateLifeOpsXPostRequest,
+  ): Promise<LifeOpsXPostResponse> {
+    return this.xDomain.createXPost(request);
+  }
+
+  getXDmDigest(opts?: {
+    accountId?: string;
+    limit?: number;
+    conversationId?: string;
   }): Promise<{
-    now: string;
-    reminderAttempts: LifeOpsReminderAttempt[];
-    workflowRuns: LifeOpsWorkflowRun[];
-    scheduledTaskFires: Array<Record<string, unknown>>;
-    scheduledTaskCompletionTimeouts: Array<Record<string, unknown>>;
-  }>;
+    generatedAt: string;
+    conversationId: string | null;
+    unreadCount: number;
+    readCount: number;
+    repliedCount: number;
+    recent: LifeOpsXDm[];
+  }> {
+    return this.xDomain.getXDmDigest(opts);
+  }
+
+  curateXDms(request: {
+    messageIds?: string[];
+    conversationId?: string;
+    markRead?: boolean;
+    markReplied?: boolean;
+  }): Promise<{ curated: number }> {
+    return this.xDomain.curateXDms(request);
+  }
+
+  sendXDirectMessage(request: {
+    participantId: string;
+    text: string;
+    confirmSend?: boolean;
+    mode?: LifeOpsConnectorMode;
+    side?: LifeOpsConnectorSide;
+    accountId?: string;
+  }): Promise<{ ok: boolean; status: number | null; error?: string }> {
+    return this.xDomain.sendXDirectMessage(request);
+  }
+
+  sendXConversationMessage(request: {
+    conversationId: string;
+    text: string;
+    confirmSend?: boolean;
+    mode?: LifeOpsConnectorMode;
+    side?: LifeOpsConnectorSide;
+    accountId?: string;
+  }): Promise<{ ok: boolean; status: number | null; error?: string }> {
+    return this.xDomain.sendXConversationMessage(request);
+  }
+
+  createXDirectMessageGroup(request: {
+    participantIds: string[];
+    text: string;
+    confirmSend?: boolean;
+    mode?: LifeOpsConnectorMode;
+    side?: LifeOpsConnectorSide;
+    accountId?: string;
+  }): Promise<{
+    ok: boolean;
+    status: number | null;
+    conversationId: string | null;
+    error?: string;
+  }> {
+    return this.xDomain.createXDirectMessageGroup(request);
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly relationshipsDomain = new RelationshipsDomain(this);
+
+  upsertRelationship(
+    input: Omit<
+      LifeOpsRelationship,
+      "id" | "agentId" | "createdAt" | "updatedAt"
+    > & { id?: string },
+  ): Promise<LifeOpsRelationship> {
+    return this.relationshipsDomain.upsertRelationship(input);
+  }
+
+  getRelationship(id: string): Promise<LifeOpsRelationship | null> {
+    return this.relationshipsDomain.getRelationship(id);
+  }
+
+  listRelationships(opts?: {
+    limit?: number;
+    primaryChannel?: LifeOpsMessageChannel;
+  }): Promise<LifeOpsRelationship[]> {
+    return this.relationshipsDomain.listRelationships(opts);
+  }
+
+  logInteraction(
+    input: Omit<LifeOpsRelationshipInteraction, "id" | "agentId" | "createdAt">,
+  ): Promise<LifeOpsRelationshipInteraction> {
+    return this.relationshipsDomain.logInteraction(input);
+  }
+
+  getDaysSinceContact(relationshipId: string): Promise<number | null> {
+    return this.relationshipsDomain.getDaysSinceContact(relationshipId);
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly emailUnsubscribeDomain = new EmailUnsubscribeDomain(this);
+
+  async scanEmailSubscriptions(
+    requestUrl: URL,
+    request: EmailUnsubscribeScanRequest = {},
+  ): Promise<EmailSubscriptionScanResult> {
+    return this.emailUnsubscribeDomain.scanEmailSubscriptions(
+      requestUrl,
+      request,
+    );
+  }
+
+  async unsubscribeEmailSender(
+    requestUrl: URL,
+    request: EmailUnsubscribeRequest,
+  ): Promise<EmailUnsubscribeResult> {
+    return this.emailUnsubscribeDomain.unsubscribeEmailSender(
+      requestUrl,
+      request,
+    );
+  }
+
+  async listEmailUnsubscribes(limit = 100): Promise<EmailUnsubscribeRecord[]> {
+    return this.emailUnsubscribeDomain.listEmailUnsubscribes(limit);
+  }
+
+  summarizeEmailUnsubscribeScan(result: EmailSubscriptionScanResult): string {
+    return this.emailUnsubscribeDomain.summarizeEmailUnsubscribeScan(result);
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly healthDomain = new HealthDomain(this);
+
+  getHealthConnectorStatus(): Promise<{
+    available: boolean;
+    backend: HealthBackend;
+    lastCheckedAt: string;
+  }> {
+    return this.healthDomain.getHealthConnectorStatus();
+  }
+
+  getHealthDataConnectorStatuses(
+    requestUrl: URL,
+    requestedMode?: LifeOpsConnectorMode,
+    requestedSide?: LifeOpsConnectorSide,
+  ): Promise<LifeOpsHealthConnectorStatus[]> {
+    return this.healthDomain.getHealthDataConnectorStatuses(
+      requestUrl,
+      requestedMode,
+      requestedSide,
+    );
+  }
+
+  getHealthDataConnectorStatus(
+    provider: LifeOpsHealthConnectorProvider,
+    requestUrl: URL,
+    requestedMode?: LifeOpsConnectorMode,
+    requestedSide?: LifeOpsConnectorSide,
+  ): Promise<LifeOpsHealthConnectorStatus> {
+    return this.healthDomain.getHealthDataConnectorStatus(
+      provider,
+      requestUrl,
+      requestedMode,
+      requestedSide,
+    );
+  }
+
+  startHealthConnector(
+    request: StartLifeOpsHealthConnectorRequest,
+    requestUrl: URL,
+  ): Promise<StartLifeOpsHealthConnectorResponse> {
+    return this.healthDomain.startHealthConnector(request, requestUrl);
+  }
+
+  completeHealthConnectorCallback(
+    callbackUrl: URL,
+  ): Promise<LifeOpsHealthConnectorStatus> {
+    return this.healthDomain.completeHealthConnectorCallback(callbackUrl);
+  }
+
+  disconnectHealthConnector(
+    request: DisconnectLifeOpsHealthConnectorRequest,
+    requestUrl: URL,
+  ): Promise<LifeOpsHealthConnectorStatus> {
+    return this.healthDomain.disconnectHealthConnector(request, requestUrl);
+  }
+
+  syncHealthConnectors(
+    request?: SyncLifeOpsHealthConnectorRequest,
+  ): Promise<LifeOpsHealthSummaryResponse> {
+    return this.healthDomain.syncHealthConnectors(request);
+  }
+
+  getHealthSummary(
+    request?: GetLifeOpsHealthSummaryRequest,
+  ): Promise<LifeOpsHealthSummaryResponse> {
+    return this.healthDomain.getHealthSummary(request);
+  }
+
+  getHealthDailySummary(date: string): Promise<HealthDailySummary> {
+    return this.healthDomain.getHealthDailySummary(date);
+  }
+
+  getHealthTrend(days: number): Promise<HealthDailySummary[]> {
+    return this.healthDomain.getHealthTrend(days);
+  }
+
+  getHealthDataPoints(opts: {
+    metric: HealthDataPoint["metric"];
+    startAt: string;
+    endAt: string;
+  }): Promise<HealthDataPoint[]> {
+    return this.healthDomain.getHealthDataPoints(opts);
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly xReadDomain = new XReadDomain(this, {
+    resolveXGrant: () => {
+      const resolver = (this as OptionalXGrantResolver).resolveXGrant;
+      return typeof resolver === "function"
+        ? resolver.call(this)
+        : Promise.resolve(null);
+    },
+  });
+
+  syncXDms(opts?: XReadOpts): Promise<{ synced: number }> {
+    return this.xReadDomain.syncXDms(opts);
+  }
+
+  syncXFeed(
+    feedType: LifeOpsXFeedType,
+    opts?: XFeedReadOpts,
+  ): Promise<{ synced: number }> {
+    return this.xReadDomain.syncXFeed(feedType, opts);
+  }
+
+  searchXPosts(query: string, opts?: XReadOpts): Promise<LifeOpsXFeedItem[]> {
+    return this.xReadDomain.searchXPosts(query, opts);
+  }
+
+  getXDms(opts?: {
+    conversationId?: string;
+    limit?: number;
+  }): Promise<LifeOpsXDm[]> {
+    return this.xReadDomain.getXDms(opts);
+  }
+
+  getXFeedItems(
+    feedType: LifeOpsXFeedType,
+    opts?: { limit?: number },
+  ): Promise<LifeOpsXFeedItem[]> {
+    return this.xReadDomain.getXFeedItems(feedType, opts);
+  }
+
+  readXInboundDms(opts?: { limit?: number }): Promise<LifeOpsXDm[]> {
+    return this.xReadDomain.readXInboundDms(opts);
+  }
+
+  // `this` satisfies LifeOpsContext. Public to avoid TS4094 on the
+  // re-exported mixin class.
+  readonly imessageDomain = new IMessageDomain(this);
+
+  getIMessageConnectorStatus(): Promise<LifeOpsIMessageConnectorStatus> {
+    return this.imessageDomain.getIMessageConnectorStatus();
+  }
+
+  sendIMessage(
+    req: IMessageSendRequest,
+  ): Promise<{ ok: true; messageId?: string }> {
+    return this.imessageDomain.sendIMessage(req);
+  }
+
+  readIMessages(opts: {
+    chatId?: string;
+    since?: string;
+    limit?: number;
+  }): Promise<IMessageRecord[]> {
+    return this.imessageDomain.readIMessages(opts);
+  }
+
+  listIMessageChats(): Promise<IMessageChat[]> {
+    return this.imessageDomain.listIMessageChats();
+  }
+
+  searchIMessages(opts: {
+    query: string;
+    chatId?: string;
+    limit?: number;
+  }): Promise<IMessageRecord[]> {
+    return this.imessageDomain.searchIMessages(opts);
+  }
+
+  getIMessageDeliveryStatus(
+    messageIds: string[],
+  ): Promise<IMessageDeliveryResult[]> {
+    return this.imessageDomain.getIMessageDeliveryStatus(messageIds);
+  }
+
+  // `this` satisfies LifeOpsContext. Public to avoid TS4094 on the
+  // re-exported mixin class.
+  readonly telegramDomain = new TelegramDomain(this);
+
+  getTelegramConnectorStatus(
+    requestedSide?: LifeOpsConnectorSide,
+  ): Promise<LifeOpsTelegramConnectorStatus> {
+    return this.telegramDomain.getTelegramConnectorStatus(requestedSide);
+  }
+
+  sendTelegramMessage(request: {
+    side?: LifeOpsConnectorSide;
+    target: string;
+    message: string;
+  }): Promise<{ ok: true; messageId: string | null }> {
+    return this.telegramDomain.sendTelegramMessage(request);
+  }
+
+  verifyTelegramConnector(
+    request: VerifyLifeOpsTelegramConnectorRequest,
+  ): Promise<VerifyLifeOpsTelegramConnectorResponse> {
+    return this.telegramDomain.verifyTelegramConnector(request);
+  }
+
+  searchTelegramMessages(request: {
+    side?: LifeOpsConnectorSide;
+    query: string;
+    scope?: string;
+    limit?: number;
+  }): Promise<TelegramMessageSearchResult[]> {
+    return this.telegramDomain.searchTelegramMessages(request);
+  }
+
+  getTelegramDeliveryStatus(_request: {
+    side?: LifeOpsConnectorSide;
+    target: string;
+    messageIds: string[];
+  }): Promise<TelegramReadReceiptResult[]> {
+    return this.telegramDomain.getTelegramDeliveryStatus(_request);
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly discordDomain = new DiscordDomain(this, {
+    createBrowserSession: (...args) => this.createBrowserSession(...args),
+    getBrowserSession: (...args) => this.getBrowserSession(...args),
+    getBrowserSettings: (...args) => this.getBrowserSettings(...args),
+    getCurrentBrowserPage: (...args) => this.getCurrentBrowserPage(...args),
+    listBrowserCompanions: (...args) => this.listBrowserCompanions(...args),
+    listBrowserTabs: (...args) => this.listBrowserTabs(...args),
+    isBrowserPaused: (...args) => this.isBrowserPaused(...args),
+  });
+
+  getDiscordConnectorStatus(
+    side?: LifeOpsConnectorSide,
+  ): Promise<LifeOpsDiscordConnectorStatus> {
+    return this.discordDomain.getDiscordConnectorStatus(side);
+  }
+
+  authorizeDiscordConnector(
+    side?: LifeOpsConnectorSide,
+    source?: LifeOpsOwnerBrowserAccessSource,
+  ): Promise<LifeOpsDiscordConnectorStatus> {
+    return this.discordDomain.authorizeDiscordConnector(side, source);
+  }
+
+  searchDiscordMessages(request: {
+    side?: LifeOpsConnectorSide;
+    query: string;
+    channelId?: string;
+    limit?: number;
+  }): Promise<DiscordMessageSearchResult[]> {
+    return this.discordDomain.searchDiscordMessages(request);
+  }
+
+  captureDiscordDeliveryStatus(
+    side?: LifeOpsConnectorSide,
+  ): Promise<DiscordMessageSearchResult[]> {
+    return this.discordDomain.captureDiscordDeliveryStatus(side);
+  }
+
+  sendDiscordMessage(request: {
+    side?: LifeOpsConnectorSide;
+    channelId?: string;
+    text: string;
+  }): Promise<DiscordSendMessageResult> {
+    return this.discordDomain.sendDiscordMessage(request);
+  }
+
+  verifyDiscordConnector(request: {
+    side?: LifeOpsConnectorSide;
+    channelId?: string;
+    sendMessage?: string;
+  }): Promise<DiscordConnectorVerification> {
+    return this.discordDomain.verifyDiscordConnector(request);
+  }
+
+  disconnectDiscord(
+    side?: LifeOpsConnectorSide,
+  ): Promise<LifeOpsDiscordConnectorStatus> {
+    return this.discordDomain.disconnectDiscord(side);
+  }
+
+  // `this` satisfies LifeOpsContext. Public to avoid TS4094 on the
+  // re-exported mixin class.
+  readonly signalDomain = new SignalDomain(this);
+
+  lifeOpsSignalServiceConnected(): boolean {
+    return this.signalDomain.lifeOpsSignalServiceConnected();
+  }
+
+  lifeOpsSignalServiceRegistered(): boolean {
+    return this.signalDomain.lifeOpsSignalServiceRegistered();
+  }
+
+  getSignalConnectorStatus(
+    side?: LifeOpsConnectorSide,
+  ): Promise<LifeOpsSignalConnectorStatus> {
+    return this.signalDomain.getSignalConnectorStatus(side);
+  }
+
+  readSignalInbound(
+    limit = 25,
+    side?: LifeOpsConnectorSide,
+  ): Promise<LifeOpsSignalInboundMessage[]> {
+    return this.signalDomain.readSignalInbound(limit, side);
+  }
+
+  sendSignalMessage(request: {
+    side?: LifeOpsConnectorSide;
+    recipient: string;
+    text: string;
+  }): Promise<{
+    provider: "signal";
+    side: LifeOpsConnectorSide;
+    recipient: string;
+    ok: true;
+    timestamp: number;
+  }> {
+    return this.signalDomain.sendSignalMessage(request);
+  }
+
+  // `this` satisfies LifeOpsContext. Public to avoid TS4094 on the
+  // re-exported mixin class.
+  readonly whatsappDomain = new WhatsAppDomain(this);
+
+  getWhatsAppConnectorStatus(): Promise<LifeOpsWhatsAppConnectorStatus> {
+    return this.whatsappDomain.getWhatsAppConnectorStatus();
+  }
+
+  sendWhatsAppMessage(
+    req: WhatsAppSendRequest,
+  ): Promise<{ ok: true; messageId: string }> {
+    return this.whatsappDomain.sendWhatsAppMessage(req);
+  }
+
+  ingestWhatsAppWebhook(
+    payload: unknown,
+  ): Promise<{ ingested: number; messages: WhatsAppMessage[] }> {
+    return this.whatsappDomain.ingestWhatsAppWebhook(payload);
+  }
+
+  syncWhatsAppInbound(): Promise<{
+    drained: number;
+    messages: WhatsAppMessage[];
+  }> {
+    return this.whatsappDomain.syncWhatsAppInbound();
+  }
+
+  pullWhatsAppRecent(limit = 25): Promise<{
+    count: number;
+    messages: WhatsAppMessage[];
+  }> {
+    return this.whatsappDomain.pullWhatsAppRecent(limit);
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly travelDomain = new TravelDomain(this, {
+    createCalendarEvent: (...args) => this.createCalendarEvent(...args),
+  });
+
+  getTravelConnectorStatus() {
+    return this.travelDomain.getTravelConnectorStatus();
+  }
+
+  searchFlights(request: SearchFlightsRequest): Promise<SearchFlightsResult> {
+    return this.travelDomain.searchFlights(request);
+  }
+
+  getFlightOffer(offerId: string): Promise<DuffelOffer> {
+    return this.travelDomain.getFlightOffer(offerId);
+  }
+
+  prepareFlightBooking(args: {
+    offerId?: string | null;
+    search?: SearchFlightsRequest | null;
+    passengers: ReadonlyArray<TravelBookingPassenger>;
+    calendarSync?: TravelCalendarSyncPlan | null;
+  }): Promise<PreparedFlightBooking> {
+    return this.travelDomain.prepareFlightBooking(args);
+  }
+
+  createFlightOrder(args: {
+    offer: DuffelOffer;
+    passengers: ReadonlyArray<TravelBookingPassenger>;
+    orderType: "hold" | "instant";
+  }): Promise<DuffelOrder> {
+    return this.travelDomain.createFlightOrder(args);
+  }
+
+  getTravelOrder(orderId: string): Promise<DuffelOrder> {
+    return this.travelDomain.getTravelOrder(orderId);
+  }
+
+  payTravelOrder(args: {
+    orderId: string;
+    amount: string;
+    currency: string;
+  }): Promise<DuffelPayment> {
+    return this.travelDomain.payTravelOrder(args);
+  }
+
+  bookFlightItinerary(
+    requestUrl: URL,
+    args: {
+      offerId?: string | null;
+      search?: SearchFlightsRequest | null;
+      passengers: ReadonlyArray<TravelBookingPassenger>;
+      calendarSync?: TravelCalendarSyncPlan | null;
+    },
+  ): Promise<FlightBookingExecutionResult> {
+    return this.travelDomain.bookFlightItinerary(requestUrl, args);
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly schedulingDomain = new SchedulingDomain(this, {
+    sendGmailMessage: (...args) => this.sendGmailMessage(...args),
+    sendTelegramMessage: (...args) => this.sendTelegramMessage(...args),
+    sendWhatsAppMessage: (...args) => this.sendWhatsAppMessage(...args),
+    sendIMessage: (...args) => this.sendIMessage(...args),
+  });
+
+  inspectSchedule(args: {
+    timezone: string;
+    now?: Date;
+  }): Promise<LifeOpsScheduleInspection> {
+    return this.schedulingDomain.inspectSchedule(args);
+  }
+
+  readScheduleSummary(args: {
+    timezone: string;
+    now?: Date;
+  }): Promise<LifeOpsScheduleSummary> {
+    return this.schedulingDomain.readScheduleSummary(args);
+  }
+
+  resolveCounterpartyTarget(
+    negotiation: LifeOpsSchedulingNegotiation,
+  ): ReturnType<SchedulingDomain["resolveCounterpartyTarget"]> {
+    return this.schedulingDomain.resolveCounterpartyTarget(negotiation);
+  }
+
+  dispatchSchedulingMessage(
+    negotiation: LifeOpsSchedulingNegotiation,
+    body: string,
+    subject: string,
+  ): ReturnType<SchedulingDomain["dispatchSchedulingMessage"]> {
+    return this.schedulingDomain.dispatchSchedulingMessage(
+      negotiation,
+      body,
+      subject,
+    );
+  }
+
+  startNegotiation(input: {
+    subject: string;
+    relationshipId?: string | null;
+    durationMinutes?: number;
+    timezone?: string;
+    metadata?: Record<string, unknown>;
+  }): Promise<LifeOpsSchedulingNegotiation> {
+    return this.schedulingDomain.startNegotiation(input);
+  }
+
+  getNegotiation(id: string): Promise<LifeOpsSchedulingNegotiation | null> {
+    return this.schedulingDomain.getNegotiation(id);
+  }
+
+  listActiveNegotiations(opts?: {
+    limit?: number;
+  }): Promise<LifeOpsSchedulingNegotiation[]> {
+    return this.schedulingDomain.listActiveNegotiations(opts);
+  }
+
+  proposeTime(input: {
+    negotiationId: string;
+    startAt: string;
+    endAt: string;
+    proposedBy: "agent" | "owner" | "counterparty";
+    metadata?: Record<string, unknown>;
+  }): Promise<LifeOpsSchedulingProposal> {
+    return this.schedulingDomain.proposeTime(input);
+  }
+
+  respondToProposal(
+    proposalId: string,
+    status: "accepted" | "declined" | "expired",
+  ): Promise<LifeOpsSchedulingProposal> {
+    return this.schedulingDomain.respondToProposal(proposalId, status);
+  }
+
+  finalizeNegotiation(
+    id: string,
+    acceptedProposalId: string,
+  ): Promise<LifeOpsSchedulingNegotiation> {
+    return this.schedulingDomain.finalizeNegotiation(id, acceptedProposalId);
+  }
+
+  cancelNegotiation(id: string, reason?: string): Promise<void> {
+    return this.schedulingDomain.cancelNegotiation(id, reason);
+  }
+
+  listProposals(negotiationId: string): Promise<LifeOpsSchedulingProposal[]> {
+    return this.schedulingDomain.listProposals(negotiationId);
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly subscriptionsDomain = new SubscriptionsDomain(this);
+
+  async listSubscriptionPlaybooks(): Promise<LifeOpsSubscriptionPlaybook[]> {
+    return this.subscriptionsDomain.listSubscriptionPlaybooks();
+  }
+
+  findSubscriptionPlaybookForMerchant(merchant: string): {
+    key: string;
+    serviceName: string;
+    managementUrl: string;
+    executorPreference: LifeOpsSubscriptionPlaybook["executorPreference"];
+  } | null {
+    return this.subscriptionsDomain.findSubscriptionPlaybookForMerchant(
+      merchant,
+    );
+  }
+
+  async getLatestSubscriptionAudit(): Promise<LifeOpsSubscriptionAuditSummary | null> {
+    return this.subscriptionsDomain.getLatestSubscriptionAudit();
+  }
+
+  async auditSubscriptions(
+    requestUrl: URL,
+    request: LifeOpsSubscriptionDiscoveryRequest = {},
+  ): Promise<LifeOpsSubscriptionAuditSummary> {
+    return this.subscriptionsDomain.auditSubscriptions(requestUrl, request);
+  }
+
+  async getSubscriptionCancellationStatus(args: {
+    cancellationId?: string | null;
+    serviceName?: string | null;
+    serviceSlug?: string | null;
+  }): Promise<LifeOpsSubscriptionCancellationSummary | null> {
+    return this.subscriptionsDomain.getSubscriptionCancellationStatus(args);
+  }
+
+  async cancelSubscription(
+    request: LifeOpsSubscriptionCancellationRequest,
+  ): Promise<LifeOpsSubscriptionCancellationSummary> {
+    return this.subscriptionsDomain.cancelSubscription(request);
+  }
+
+  summarizeSubscriptionAudit(summary: LifeOpsSubscriptionAuditSummary): string {
+    return this.subscriptionsDomain.summarizeSubscriptionAudit(summary);
+  }
+
+  summarizeSubscriptionCancellation(
+    summary: LifeOpsSubscriptionCancellationSummary,
+  ): string {
+    return this.subscriptionsDomain.summarizeSubscriptionCancellation(summary);
+  }
+
+  resolveSubscriptionIntent(text: string): {
+    mode: "audit" | "cancel" | "status" | null;
+    serviceName?: string;
+    serviceSlug?: string;
+    executor?: LifeOpsSubscriptionExecutor;
+  } {
+    return this.subscriptionsDomain.resolveSubscriptionIntent(text);
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly statusDomain = new StatusDomain(this, {
+    getScheduleMergedState: (...args) => this.getScheduleMergedState(...args),
+    getBrowserSettings: (...args) => this.getBrowserSettings(...args),
+    listBrowserCompanions: (...args) => this.listBrowserCompanions(...args),
+    getXConnectorStatus: (...args) => this.getXConnectorStatus(...args),
+    getHealthConnectorStatus: (...args) =>
+      this.getHealthConnectorStatus(...args),
+  });
+
+  getCapabilityStatus(now?: Date): Promise<LifeOpsCapabilitiesStatus> {
+    return this.statusDomain.getCapabilityStatus(now);
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly screenTimeDomain = new ScreenTimeDomain(this, {
+    getBrowserSettings: (...args) => this.getBrowserSettings(...args),
+    listBrowserCompanions: (...args) => this.listBrowserCompanions(...args),
+  });
+
+  recordScreenTimeEvent(
+    event: ScreenTimeEventInput,
+  ): Promise<LifeOpsScreenTimeSession> {
+    return this.screenTimeDomain.recordScreenTimeEvent(event);
+  }
+
+  finishActiveScreenTimeSession(
+    id: string,
+    endAt: string,
+    durationSeconds: number,
+  ): Promise<void> {
+    return this.screenTimeDomain.finishActiveScreenTimeSession(
+      id,
+      endAt,
+      durationSeconds,
+    );
+  }
+
+  collectScreenTimeRows(opts: {
+    since: string;
+    until: string;
+    source?: LifeOpsScreenTimeSource;
+    identifier?: string;
+  }): Promise<ScreenTimeAggregateRow[]> {
+    return this.screenTimeDomain.collectScreenTimeRows(opts);
+  }
+
+  getScreenTimeDaily(opts: {
+    date: string;
+    source?: LifeOpsScreenTimeSource;
+    identifier?: string;
+    limit?: number;
+  }): Promise<LifeOpsScreenTimeDaily[]> {
+    return this.screenTimeDomain.getScreenTimeDaily(opts);
+  }
+
+  getScreenTimeSummary(opts: {
+    since: string;
+    until: string;
+    source?: LifeOpsScreenTimeSource;
+    identifier?: string;
+    topN?: number;
+  }): Promise<LifeOpsScreenTimeSummary> {
+    return this.screenTimeDomain.getScreenTimeSummary(opts);
+  }
+
+  getScreenTimeBreakdown(opts: {
+    since: string;
+    until: string;
+    source?: LifeOpsScreenTimeSource;
+    identifier?: string;
+    topN?: number;
+  }): Promise<ScreenTimeBreakdown> {
+    return this.screenTimeDomain.getScreenTimeBreakdown(opts);
+  }
+
+  getSocialHabitSummary(opts: {
+    since: string;
+    until: string;
+    topN?: number;
+  }): Promise<SocialHabitSummary> {
+    return this.screenTimeDomain.getSocialHabitSummary(opts);
+  }
+
+  getScreenTimeHistory(opts: {
+    range: LifeOpsScreenTimeRangeKey;
+    topN?: number;
+    socialTopN?: number;
+  }): Promise<LifeOpsScreenTimeHistoryResponse> {
+    return this.screenTimeDomain.getScreenTimeHistory(opts);
+  }
+
+  getScreenTimeWeeklyAverageByApp(opts: {
+    since: string;
+    until: string;
+    daysInWindow: number;
+    identifier?: string;
+    topN?: number;
+  }): Promise<ScreenTimeWeeklyAverageResponse> {
+    return this.screenTimeDomain.getScreenTimeWeeklyAverageByApp(opts);
+  }
+
+  aggregateDailyForDate(date: string): Promise<{ updated: number }> {
+    return this.screenTimeDomain.aggregateDailyForDate(date);
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly sleepDomain = new SleepDomain(this);
+
+  getSleepHistory(opts?: {
+    windowDays?: number;
+    includeNaps?: boolean;
+  }): Promise<LifeOpsSleepHistoryResponse> {
+    return this.sleepDomain.getSleepHistory(opts);
+  }
+
+  getSleepRegularity(opts?: {
+    windowDays?: number;
+    includeNaps?: boolean;
+  }): Promise<LifeOpsSleepRegularityResponse> {
+    return this.sleepDomain.getSleepRegularity(opts);
+  }
+
+  getPersonalBaseline(opts?: {
+    windowDays?: number;
+  }): Promise<LifeOpsPersonalBaselineResponse> {
+    return this.sleepDomain.getPersonalBaseline(opts);
+  }
+
+  // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+  // Public (not private) to avoid TS4094 on the re-exported mixin class.
+  readonly inboxDomain = new InboxDomain(this, {
+    getGoogleConnectorStatus: (...args) =>
+      this.getGoogleConnectorStatus(...args),
+    getGmailTriage: (...args) => this.getGmailTriage(...args),
+    getXConnectorStatus: (...args) => this.getXConnectorStatus(...args),
+    syncXDms: (...args) => this.syncXDms(...args),
+    getXDms: (...args) => this.getXDms(...args),
+  });
+
+  getInbox(request: GetLifeOpsInboxRequest = {}): Promise<LifeOpsInbox> {
+    return this.inboxDomain.getInbox(request);
+  }
+
+  markInboxEntryRead(inboxEntryId: string): Promise<LifeOpsInboxMessage> {
+    return this.inboxDomain.markInboxEntryRead(inboxEntryId);
+  }
 }
